% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_event_type.R
\docType{class}
\name{ActivityEventType}
\alias{ActivityEventType}
\title{ActivityEventType}
\format{
An \code{R6Class} generator object
}
\description{
ActivityEventType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ActivityEventType-new}{\code{ActivityEventType$new()}}
\item \href{#method-ActivityEventType-toJSON}{\code{ActivityEventType$toJSON()}}
\item \href{#method-ActivityEventType-fromJSON}{\code{ActivityEventType$fromJSON()}}
\item \href{#method-ActivityEventType-toJSONString}{\code{ActivityEventType$toJSONString()}}
\item \href{#method-ActivityEventType-fromJSONString}{\code{ActivityEventType$fromJSONString()}}
\item \href{#method-ActivityEventType-clone}{\code{ActivityEventType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityEventType-new"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityEventType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventType$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityEventType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityEventType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityEventType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityEventType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventType$fromJSON(ActivityEventTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityEventType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityEventType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityEventType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityEventType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventType$fromJSONString(ActivityEventTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ActivityEventType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ActivityEventType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ActivityEventType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
