% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\docType{class}
\name{Array}
\alias{Array}
\title{Array}
\format{
An \code{R6Class} generator object
}
\description{
Array Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{timestamp}}{numeric}

\item{\code{queryType}}{\link{Querytype}}

\item{\code{uri}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Array-new}{\code{Array$new()}}
\item \href{#method-Array-toJSON}{\code{Array$toJSON()}}
\item \href{#method-Array-fromJSON}{\code{Array$fromJSON()}}
\item \href{#method-Array-toJSONString}{\code{Array$toJSONString()}}
\item \href{#method-Array-fromJSONString}{\code{Array$fromJSONString()}}
\item \href{#method-Array-clone}{\code{Array$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Array-new"></a>}}
\if{latex}{\out{\hypertarget{method-Array-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Array$new(timestamp, queryType, uri, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Array-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Array-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Array$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Array-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Array-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Array$fromJSON(ArrayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Array-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Array-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Array$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Array-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Array-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Array$fromJSONString(ArrayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Array-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Array-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Array$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
