% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_actions.R
\docType{class}
\name{ArrayActions}
\alias{ArrayActions}
\title{ArrayActions}
\format{
An \code{R6Class} generator object
}
\description{
ArrayActions Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArrayActions-new}{\code{ArrayActions$new()}}
\item \href{#method-ArrayActions-toJSON}{\code{ArrayActions$toJSON()}}
\item \href{#method-ArrayActions-fromJSON}{\code{ArrayActions$fromJSON()}}
\item \href{#method-ArrayActions-toJSONString}{\code{ArrayActions$toJSONString()}}
\item \href{#method-ArrayActions-fromJSONString}{\code{ArrayActions$fromJSONString()}}
\item \href{#method-ArrayActions-clone}{\code{ArrayActions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayActions-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayActions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayActions$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayActions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayActions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayActions$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayActions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayActions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayActions$fromJSON(ArrayActionsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayActions-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayActions-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayActions$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayActions-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayActions-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayActions$fromJSONString(ArrayActionsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayActions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayActions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayActions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
