% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_info_update.R
\docType{class}
\name{ArrayInfoUpdate}
\alias{ArrayInfoUpdate}
\title{ArrayInfoUpdate}
\format{
An \code{R6Class} generator object
}
\description{
ArrayInfoUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{uri}}{character [optional]}

\item{\code{file_type}}{\link{FileType} [optional]}

\item{\code{file_properties}}{named list( character ) [optional]}

\item{\code{access_credentials_name}}{character [optional]}

\item{\code{logo}}{character [optional]}

\item{\code{tags}}{list( character ) [optional]}

\item{\code{license_id}}{character [optional]}

\item{\code{license_text}}{character [optional]}

\item{\code{read_only}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArrayInfoUpdate-new}{\code{ArrayInfoUpdate$new()}}
\item \href{#method-ArrayInfoUpdate-toJSON}{\code{ArrayInfoUpdate$toJSON()}}
\item \href{#method-ArrayInfoUpdate-fromJSON}{\code{ArrayInfoUpdate$fromJSON()}}
\item \href{#method-ArrayInfoUpdate-toJSONString}{\code{ArrayInfoUpdate$toJSONString()}}
\item \href{#method-ArrayInfoUpdate-fromJSONString}{\code{ArrayInfoUpdate$fromJSONString()}}
\item \href{#method-ArrayInfoUpdate-clone}{\code{ArrayInfoUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayInfoUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayInfoUpdate-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayInfoUpdate$new(
  description = NULL,
  name = NULL,
  uri = NULL,
  file_type = NULL,
  file_properties = NULL,
  access_credentials_name = NULL,
  logo = NULL,
  tags = NULL,
  license_id = NULL,
  license_text = NULL,
  read_only = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayInfoUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayInfoUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayInfoUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayInfoUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayInfoUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayInfoUpdate$fromJSON(ArrayInfoUpdateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayInfoUpdate-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayInfoUpdate-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayInfoUpdate$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayInfoUpdate-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayInfoUpdate-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayInfoUpdate$fromJSONString(ArrayInfoUpdateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayInfoUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayInfoUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayInfoUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
