% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_sharing.R
\docType{class}
\name{ArraySharing}
\alias{ArraySharing}
\title{ArraySharing}
\format{
An \code{R6Class} generator object
}
\description{
ArraySharing Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actions}}{list( \link{ArrayActions} ) [optional]}

\item{\code{namespace}}{character [optional]}

\item{\code{namespace_type}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArraySharing-new}{\code{ArraySharing$new()}}
\item \href{#method-ArraySharing-toJSON}{\code{ArraySharing$toJSON()}}
\item \href{#method-ArraySharing-fromJSON}{\code{ArraySharing$fromJSON()}}
\item \href{#method-ArraySharing-toJSONString}{\code{ArraySharing$toJSONString()}}
\item \href{#method-ArraySharing-fromJSONString}{\code{ArraySharing$fromJSONString()}}
\item \href{#method-ArraySharing-clone}{\code{ArraySharing$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraySharing-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArraySharing-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArraySharing$new(actions = NULL, namespace = NULL, namespace_type = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraySharing-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArraySharing-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArraySharing$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraySharing-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArraySharing-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArraySharing$fromJSON(ArraySharingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraySharing-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArraySharing-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArraySharing$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraySharing-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArraySharing-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArraySharing$fromJSONString(ArraySharingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArraySharing-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArraySharing-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArraySharing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
