% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_task.R
\docType{class}
\name{ArrayTask}
\alias{ArrayTask}
\title{ArrayTask}
\format{
An \code{R6Class} generator object
}
\description{
ArrayTask Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{array_metadata}}{\link{ArrayInfo} [optional]}

\item{\code{subarray}}{\link{DomainArray} [optional]}

\item{\code{memory}}{integer [optional]}

\item{\code{cpu}}{integer [optional]}

\item{\code{namespace}}{character [optional]}

\item{\code{status}}{\link{ArrayTaskStatus} [optional]}

\item{\code{start_time}}{character [optional]}

\item{\code{finish_time}}{character [optional]}

\item{\code{cost}}{numeric [optional]}

\item{\code{egress_cost}}{numeric [optional]}

\item{\code{access_cost}}{numeric [optional]}

\item{\code{query_type}}{\link{Querytype} [optional]}

\item{\code{udf_code}}{character [optional]}

\item{\code{udf_language}}{character [optional]}

\item{\code{sql_query}}{character [optional]}

\item{\code{type}}{\link{ArrayTaskType} [optional]}

\item{\code{activity}}{list( \link{ArrayActivityLog} ) [optional]}

\item{\code{logs}}{character [optional]}

\item{\code{duration}}{numeric [optional]}

\item{\code{sql_init_commands}}{list( character ) [optional]}

\item{\code{sql_parameters}}{list( object ) [optional]}

\item{\code{result_format}}{\link{ResultFormat} [optional]}

\item{\code{task_graph_uuid}}{character [optional]}

\item{\code{client_node_uuid}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArrayTask-new}{\code{ArrayTask$new()}}
\item \href{#method-ArrayTask-toJSON}{\code{ArrayTask$toJSON()}}
\item \href{#method-ArrayTask-fromJSON}{\code{ArrayTask$fromJSON()}}
\item \href{#method-ArrayTask-toJSONString}{\code{ArrayTask$toJSONString()}}
\item \href{#method-ArrayTask-fromJSONString}{\code{ArrayTask$fromJSONString()}}
\item \href{#method-ArrayTask-clone}{\code{ArrayTask$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTask-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTask-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTask$new(
  id = NULL,
  name = NULL,
  description = NULL,
  array_metadata = NULL,
  subarray = NULL,
  memory = NULL,
  cpu = NULL,
  namespace = NULL,
  status = NULL,
  start_time = NULL,
  finish_time = NULL,
  cost = NULL,
  egress_cost = NULL,
  access_cost = NULL,
  query_type = NULL,
  udf_code = NULL,
  udf_language = NULL,
  sql_query = NULL,
  type = NULL,
  activity = NULL,
  logs = NULL,
  duration = NULL,
  sql_init_commands = NULL,
  sql_parameters = NULL,
  result_format = NULL,
  task_graph_uuid = NULL,
  client_node_uuid = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTask-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTask-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTask$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTask-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTask-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTask$fromJSON(ArrayTaskJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTask-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTask-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTask$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTask-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTask-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTask$fromJSONString(ArrayTaskJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTask-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTask-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
