% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_task_browser_sidebar.R
\docType{class}
\name{ArrayTaskBrowserSidebar}
\alias{ArrayTaskBrowserSidebar}
\title{ArrayTaskBrowserSidebar}
\format{
An \code{R6Class} generator object
}
\description{
ArrayTaskBrowserSidebar Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{organizations}}{list( character ) [optional]}

\item{\code{result_count_for_all}}{integer [optional]}

\item{\code{result_count_by_namespace}}{object [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArrayTaskBrowserSidebar-new}{\code{ArrayTaskBrowserSidebar$new()}}
\item \href{#method-ArrayTaskBrowserSidebar-toJSON}{\code{ArrayTaskBrowserSidebar$toJSON()}}
\item \href{#method-ArrayTaskBrowserSidebar-fromJSON}{\code{ArrayTaskBrowserSidebar$fromJSON()}}
\item \href{#method-ArrayTaskBrowserSidebar-toJSONString}{\code{ArrayTaskBrowserSidebar$toJSONString()}}
\item \href{#method-ArrayTaskBrowserSidebar-fromJSONString}{\code{ArrayTaskBrowserSidebar$fromJSONString()}}
\item \href{#method-ArrayTaskBrowserSidebar-clone}{\code{ArrayTaskBrowserSidebar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskBrowserSidebar-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskBrowserSidebar-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskBrowserSidebar$new(
  organizations = NULL,
  result_count_for_all = NULL,
  result_count_by_namespace = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskBrowserSidebar-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskBrowserSidebar-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskBrowserSidebar$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskBrowserSidebar-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskBrowserSidebar-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskBrowserSidebar$fromJSON(ArrayTaskBrowserSidebarJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskBrowserSidebar-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskBrowserSidebar-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskBrowserSidebar$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskBrowserSidebar-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskBrowserSidebar-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskBrowserSidebar$fromJSONString(ArrayTaskBrowserSidebarJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskBrowserSidebar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskBrowserSidebar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskBrowserSidebar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
