% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_task_data.R
\docType{class}
\name{ArrayTaskData}
\alias{ArrayTaskData}
\title{ArrayTaskData}
\format{
An \code{R6Class} generator object
}
\description{
ArrayTaskData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{array_tasks}}{list( \link{ArrayTask} ) [optional]}

\item{\code{pagination_metadata}}{\link{PaginationMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArrayTaskData-new}{\code{ArrayTaskData$new()}}
\item \href{#method-ArrayTaskData-toJSON}{\code{ArrayTaskData$toJSON()}}
\item \href{#method-ArrayTaskData-fromJSON}{\code{ArrayTaskData$fromJSON()}}
\item \href{#method-ArrayTaskData-toJSONString}{\code{ArrayTaskData$toJSONString()}}
\item \href{#method-ArrayTaskData-fromJSONString}{\code{ArrayTaskData$fromJSONString()}}
\item \href{#method-ArrayTaskData-clone}{\code{ArrayTaskData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskData$new(array_tasks = NULL, pagination_metadata = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskData$fromJSON(ArrayTaskDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskData$fromJSONString(ArrayTaskDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
