% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_task_log.R
\docType{class}
\name{ArrayTaskLog}
\alias{ArrayTaskLog}
\title{ArrayTaskLog}
\format{
An \code{R6Class} generator object
}
\description{
ArrayTaskLog Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{array_task_id}}{character [optional]}

\item{\code{logs}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArrayTaskLog-new}{\code{ArrayTaskLog$new()}}
\item \href{#method-ArrayTaskLog-toJSON}{\code{ArrayTaskLog$toJSON()}}
\item \href{#method-ArrayTaskLog-fromJSON}{\code{ArrayTaskLog$fromJSON()}}
\item \href{#method-ArrayTaskLog-toJSONString}{\code{ArrayTaskLog$toJSONString()}}
\item \href{#method-ArrayTaskLog-fromJSONString}{\code{ArrayTaskLog$fromJSONString()}}
\item \href{#method-ArrayTaskLog-clone}{\code{ArrayTaskLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskLog-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskLog-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskLog$new(array_task_id = NULL, logs = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskLog-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskLog-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskLog$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskLog-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskLog-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskLog$fromJSON(ArrayTaskLogJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskLog-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskLog-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskLog$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskLog-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskLog-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskLog$fromJSONString(ArrayTaskLogJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskLog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskLog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
