% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_task_status.R
\docType{class}
\name{ArrayTaskStatus}
\alias{ArrayTaskStatus}
\title{ArrayTaskStatus}
\format{
An \code{R6Class} generator object
}
\description{
ArrayTaskStatus Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArrayTaskStatus-new}{\code{ArrayTaskStatus$new()}}
\item \href{#method-ArrayTaskStatus-toJSON}{\code{ArrayTaskStatus$toJSON()}}
\item \href{#method-ArrayTaskStatus-fromJSON}{\code{ArrayTaskStatus$fromJSON()}}
\item \href{#method-ArrayTaskStatus-toJSONString}{\code{ArrayTaskStatus$toJSONString()}}
\item \href{#method-ArrayTaskStatus-fromJSONString}{\code{ArrayTaskStatus$fromJSONString()}}
\item \href{#method-ArrayTaskStatus-clone}{\code{ArrayTaskStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskStatus-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskStatus$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskStatus$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskStatus$fromJSON(ArrayTaskStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskStatus$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskStatus$fromJSONString(ArrayTaskStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
