% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_task_type.R
\docType{class}
\name{ArrayTaskType}
\alias{ArrayTaskType}
\title{ArrayTaskType}
\format{
An \code{R6Class} generator object
}
\description{
ArrayTaskType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArrayTaskType-new}{\code{ArrayTaskType$new()}}
\item \href{#method-ArrayTaskType-toJSON}{\code{ArrayTaskType$toJSON()}}
\item \href{#method-ArrayTaskType-fromJSON}{\code{ArrayTaskType$fromJSON()}}
\item \href{#method-ArrayTaskType-toJSONString}{\code{ArrayTaskType$toJSONString()}}
\item \href{#method-ArrayTaskType-fromJSONString}{\code{ArrayTaskType$fromJSONString()}}
\item \href{#method-ArrayTaskType-clone}{\code{ArrayTaskType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskType-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskType$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskType$fromJSON(ArrayTaskTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskType$fromJSONString(ArrayTaskTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayTaskType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayTaskType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayTaskType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
