% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_type.R
\docType{class}
\name{ArrayType}
\alias{ArrayType}
\title{ArrayType}
\format{
An \code{R6Class} generator object
}
\description{
ArrayType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArrayType-new}{\code{ArrayType$new()}}
\item \href{#method-ArrayType-toJSON}{\code{ArrayType$toJSON()}}
\item \href{#method-ArrayType-fromJSON}{\code{ArrayType$fromJSON()}}
\item \href{#method-ArrayType-toJSONString}{\code{ArrayType$toJSONString()}}
\item \href{#method-ArrayType-fromJSONString}{\code{ArrayType$fromJSONString()}}
\item \href{#method-ArrayType-clone}{\code{ArrayType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayType-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayType$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayType$fromJSON(ArrayTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayType$fromJSONString(ArrayTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
