% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.R
\docType{class}
\name{Attribute}
\alias{Attribute}
\title{Attribute}
\format{
An \code{R6Class} generator object
}
\description{
Attribute Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character}

\item{\code{type}}{\link{Datatype}}

\item{\code{filterPipeline}}{\link{FilterPipeline}}

\item{\code{cellValNum}}{integer}

\item{\code{nullable}}{character [optional]}

\item{\code{fillValue}}{list( integer ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Attribute-new}{\code{Attribute$new()}}
\item \href{#method-Attribute-toJSON}{\code{Attribute$toJSON()}}
\item \href{#method-Attribute-fromJSON}{\code{Attribute$fromJSON()}}
\item \href{#method-Attribute-toJSONString}{\code{Attribute$toJSONString()}}
\item \href{#method-Attribute-fromJSONString}{\code{Attribute$fromJSONString()}}
\item \href{#method-Attribute-clone}{\code{Attribute$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Attribute-new"></a>}}
\if{latex}{\out{\hypertarget{method-Attribute-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Attribute$new(
  name,
  type,
  filterPipeline,
  cellValNum,
  nullable = NULL,
  fillValue = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Attribute-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Attribute-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Attribute$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Attribute-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Attribute-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Attribute$fromJSON(AttributeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Attribute-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Attribute-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Attribute$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Attribute-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Attribute-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Attribute$fromJSONString(AttributeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Attribute-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Attribute-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Attribute$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
