% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_buffer_header.R
\docType{class}
\name{AttributeBufferHeader}
\alias{AttributeBufferHeader}
\title{AttributeBufferHeader}
\format{
An \code{R6Class} generator object
}
\description{
AttributeBufferHeader Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character}

\item{\code{fixedLenBufferSizeInBytes}}{integer}

\item{\code{varLenBufferSizeInBytes}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AttributeBufferHeader-new}{\code{AttributeBufferHeader$new()}}
\item \href{#method-AttributeBufferHeader-toJSON}{\code{AttributeBufferHeader$toJSON()}}
\item \href{#method-AttributeBufferHeader-fromJSON}{\code{AttributeBufferHeader$fromJSON()}}
\item \href{#method-AttributeBufferHeader-toJSONString}{\code{AttributeBufferHeader$toJSONString()}}
\item \href{#method-AttributeBufferHeader-fromJSONString}{\code{AttributeBufferHeader$fromJSONString()}}
\item \href{#method-AttributeBufferHeader-clone}{\code{AttributeBufferHeader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeBufferHeader-new"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeBufferHeader-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeBufferHeader$new(
  name,
  fixedLenBufferSizeInBytes,
  varLenBufferSizeInBytes,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeBufferHeader-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeBufferHeader-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeBufferHeader$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeBufferHeader-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeBufferHeader-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeBufferHeader$fromJSON(AttributeBufferHeaderJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeBufferHeader-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeBufferHeader-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeBufferHeader$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeBufferHeader-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeBufferHeader-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeBufferHeader$fromJSONString(AttributeBufferHeaderJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AttributeBufferHeader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AttributeBufferHeader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AttributeBufferHeader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
