% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_access_credentials.R
\docType{class}
\name{AWSAccessCredentials}
\alias{AWSAccessCredentials}
\title{AWSAccessCredentials}
\format{
An \code{R6Class} generator object
}
\description{
AWSAccessCredentials Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{secret_access_key}}{character [optional]}

\item{\code{access_key_id}}{character [optional]}

\item{\code{service_role_arn}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{default}}{character [optional]}

\item{\code{buckets}}{list( character ) [optional]}

\item{\code{created_at}}{character [optional]}

\item{\code{updated_at}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AWSAccessCredentials-new}{\code{AWSAccessCredentials$new()}}
\item \href{#method-AWSAccessCredentials-toJSON}{\code{AWSAccessCredentials$toJSON()}}
\item \href{#method-AWSAccessCredentials-fromJSON}{\code{AWSAccessCredentials$fromJSON()}}
\item \href{#method-AWSAccessCredentials-toJSONString}{\code{AWSAccessCredentials$toJSONString()}}
\item \href{#method-AWSAccessCredentials-fromJSONString}{\code{AWSAccessCredentials$fromJSONString()}}
\item \href{#method-AWSAccessCredentials-clone}{\code{AWSAccessCredentials$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AWSAccessCredentials-new"></a>}}
\if{latex}{\out{\hypertarget{method-AWSAccessCredentials-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AWSAccessCredentials$new(
  secret_access_key = NULL,
  access_key_id = NULL,
  service_role_arn = NULL,
  name = NULL,
  default = NULL,
  buckets = NULL,
  created_at = NULL,
  updated_at = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AWSAccessCredentials-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AWSAccessCredentials-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AWSAccessCredentials$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AWSAccessCredentials-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-AWSAccessCredentials-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AWSAccessCredentials$fromJSON(AWSAccessCredentialsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AWSAccessCredentials-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AWSAccessCredentials-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AWSAccessCredentials$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AWSAccessCredentials-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-AWSAccessCredentials-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AWSAccessCredentials$fromJSONString(AWSAccessCredentialsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AWSAccessCredentials-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AWSAccessCredentials-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AWSAccessCredentials$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
