% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_node.R
\name{compute}
\alias{compute}
\title{Launch a task graph from a given terminal node in the task graph}
\usage{
compute(object, namespace, timeout_seconds = NULL, verbose = FALSE)
}
\description{
The task graph is implicitly defined by various \code{delayed} objects having others
in their argument lists.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{delayed}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
