% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_node.R
\name{compute_sequentially}
\alias{compute_sequentially}
\title{Test/debug entrypoint for local/sequential compute.}
\usage{
compute_sequentially(node)
}
\arguments{
\item{node}{Nominally, produced by \code{delayed}, \code{delayed_generic_udf}, etc.}
}
\value{
The value of the computation.
}
\description{
Runs all nodes in a correct dependency ordering, but all within the context of the same
process, and all locally. See also the Task Graphs vignette.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute}()},
\code{\link{delayed_args<-}()},
\code{\link{delayed_args}()},
\code{\link{delayed_array_udf}()},
\code{\link{delayed_generic_udf}()},
\code{\link{delayed_sql}()},
\code{\link{delayed}()},
\code{\link{deregister_array}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_array}()},
\code{\link{register_udf}()},
\code{\link{share_array}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
