% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\docType{class}
\name{Datatype}
\alias{Datatype}
\title{Datatype}
\format{
An \code{R6Class} generator object
}
\description{
Datatype Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Datatype-new}{\code{Datatype$new()}}
\item \href{#method-Datatype-toJSON}{\code{Datatype$toJSON()}}
\item \href{#method-Datatype-fromJSON}{\code{Datatype$fromJSON()}}
\item \href{#method-Datatype-toJSONString}{\code{Datatype$toJSONString()}}
\item \href{#method-Datatype-fromJSONString}{\code{Datatype$fromJSONString()}}
\item \href{#method-Datatype-clone}{\code{Datatype$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Datatype-new"></a>}}
\if{latex}{\out{\hypertarget{method-Datatype-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Datatype$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Datatype-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Datatype-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Datatype$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Datatype-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Datatype-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Datatype$fromJSON(DatatypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Datatype-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Datatype-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Datatype$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Datatype-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Datatype-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Datatype$fromJSONString(DatatypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Datatype-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Datatype-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Datatype$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
