% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_node.R
\name{delayed_args<-}
\alias{delayed_args<-}
\title{Set arguments for a delayed function.}
\usage{
delayed_args(node) <- value
}
\arguments{
\item{node}{The object whose args are being set -- nominally, produced by
\code{delayed}, \code{delayed_generic_udf}, etc.}

\item{value}{A list of arguments to the delayed function, e.g. \code{list(a,b,c)}.}
}
\description{
Args can be set when \code{delayed} is called, or afterward using this function.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute_sequentially}()},
\code{\link{compute}()},
\code{\link{delayed_args}()},
\code{\link{delayed_array_udf}()},
\code{\link{delayed_generic_udf}()},
\code{\link{delayed_sql}()},
\code{\link{delayed}()},
\code{\link{deregister_array}()},
\code{\link{deregister_group}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{list_groups}()},
\code{\link{login}()},
\code{\link{register_array}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
