% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_delayed.R
\name{delayed_generic_udf}
\alias{delayed_generic_udf}
\title{Define a generic UDF to be executed within a task graph}
\usage{
delayed_generic_udf(
  namespace,
  udf = NULL,
  registered_udf_name = NULL,
  args = NULL,
  name = NULL
)
}
\arguments{
\item{namespace}{The TileDB-Cloud namespace to charge the query to}

\item{udf}{An R function. Arguments are specified separately via \code{args}.
One of \code{udf} and \code{registered_udf_name} must be non-null.}

\item{registered_udf_name}{Name of a registered UDF, of the form \code{namespace/udfname}.
Arguments are specified separately via \code{args}.  One of \code{udf} and
\code{registered_udf_name} must be non-null.}

\item{name}{Optional -- e.g. \code{a} or \code{b}. If omitted, it defaults to a UUID.}
}
\value{
The return value from the UDF as an R object.
}
\description{
Define a generic UDF to be executed within a task graph
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute_sequentially}()},
\code{\link{compute}()},
\code{\link{delayed_args<-}()},
\code{\link{delayed_args}()},
\code{\link{delayed_array_udf}()},
\code{\link{delayed_sql}()},
\code{\link{delayed}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
