% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_delayed.R
\name{delayedArrayUDF}
\alias{delayedArrayUDF}
\title{Define a single-array UDF to be executed within a task graph}
\usage{
delayedArrayUDF(namespace, array, udf, selectedRanges, attrs)
}
\arguments{
\item{namespace}{The TileDB-Cloud namespace to charge the query to}

\item{array}{TileDB URI -- see vignette for examples.}

\item{udf}{User-defined function, as in UDF examples.}

\item{selectedRanges}{As in UDF examples.}

\item{attrs}{As in UDF examples.}
}
\value{
The return value from the UDF as an R object.
}
\description{
Define a single-array UDF to be executed within a task graph
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute}()},
\code{\link{delayedSQL}()},
\code{\link{delayed_args<-}()},
\code{\link{delayed_args}()},
\code{\link{delayed}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
