% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_group_deregister.R
\name{deregister_group}
\alias{deregister_group}
\title{De-register a 'Group' object recursively}
\usage{
deregister_group(
  uri,
  namespace,
  name,
  delete_from_group = TRUE,
  delete_array = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{uri}{A TileDB + S3 URI}

\item{namespace}{A character like "TileDB-Inc"}

\item{name}{A characther "groupABC"}

\item{delete_from_group}{A logical value, default `TRUE`, whether arrays are removed from the group}

\item{delete_array}{A logical value, default `TRUE`, whether arrays are deleted too}

\item{verbose}{A logical value, default `FALSE`, whether operations are verbose or not}
}
\value{
Nothing is returned, the function is invoked for its side-effect
}
\description{
This function de-registers a 'Group' object, the 'Group' objects therein
as well as any arrays.
}
\details{
Note that 'Group' objects remain on the underlying storage such as S3.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute_sequentially}()},
\code{\link{compute}()},
\code{\link{delayed_args<-}()},
\code{\link{delayed_args}()},
\code{\link{delayed_array_udf}()},
\code{\link{delayed_generic_udf}()},
\code{\link{delayed_sql}()},
\code{\link{delayed}()},
\code{\link{deregister_array}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{list_groups}()},
\code{\link{login}()},
\code{\link{register_array}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
