% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_udf.R
\name{deregister_udf}
\alias{deregister_udf}
\title{Deregister a UDF from TileDB Cloud}
\usage{
deregister_udf(name, namespace)
}
\arguments{
\item{name}{Name of the UDF in TileDB Cloud, e.g. \code{myudfname}.}

\item{namespace}{Namespace for the UDF in TileDB Cloud, e.g. \code{mynamespace}.}
}
\value{
No return value.
}
\description{
Deletes a registered UDF. This removes all sharing and cannot be undone.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute_sequentially}()},
\code{\link{compute}()},
\code{\link{delayed_args<-}()},
\code{\link{delayed_args}()},
\code{\link{delayed_array_udf}()},
\code{\link{delayed_generic_udf}()},
\code{\link{delayed_sql}()},
\code{\link{delayed}()},
\code{\link{deregister_array}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_array}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
