% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_coordinate.R
\docType{class}
\name{DimensionCoordinate}
\alias{DimensionCoordinate}
\title{DimensionCoordinate}
\format{
An \code{R6Class} generator object
}
\description{
DimensionCoordinate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{int8}}{integer [optional]}

\item{\code{uint8}}{integer [optional]}

\item{\code{int16}}{integer [optional]}

\item{\code{uint16}}{integer [optional]}

\item{\code{int32}}{integer [optional]}

\item{\code{uint32}}{integer [optional]}

\item{\code{int64}}{integer [optional]}

\item{\code{uint64}}{integer [optional]}

\item{\code{float32}}{numeric [optional]}

\item{\code{float64}}{numeric [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DimensionCoordinate-new}{\code{DimensionCoordinate$new()}}
\item \href{#method-DimensionCoordinate-toJSON}{\code{DimensionCoordinate$toJSON()}}
\item \href{#method-DimensionCoordinate-fromJSON}{\code{DimensionCoordinate$fromJSON()}}
\item \href{#method-DimensionCoordinate-toJSONString}{\code{DimensionCoordinate$toJSONString()}}
\item \href{#method-DimensionCoordinate-fromJSONString}{\code{DimensionCoordinate$fromJSONString()}}
\item \href{#method-DimensionCoordinate-clone}{\code{DimensionCoordinate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionCoordinate-new"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionCoordinate-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionCoordinate$new(
  int8 = NULL,
  uint8 = NULL,
  int16 = NULL,
  uint16 = NULL,
  int32 = NULL,
  uint32 = NULL,
  int64 = NULL,
  uint64 = NULL,
  float32 = NULL,
  float64 = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionCoordinate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionCoordinate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionCoordinate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionCoordinate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionCoordinate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionCoordinate$fromJSON(DimensionCoordinateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionCoordinate-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionCoordinate-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionCoordinate$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionCoordinate-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionCoordinate-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionCoordinate$fromJSONString(DimensionCoordinateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionCoordinate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionCoordinate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionCoordinate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
