% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_tile_extent.R
\docType{class}
\name{DimensionTileExtent}
\alias{DimensionTileExtent}
\title{DimensionTileExtent}
\format{
An \code{R6Class} generator object
}
\description{
DimensionTileExtent Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{int8}}{integer [optional]}

\item{\code{uint8}}{integer [optional]}

\item{\code{int16}}{integer [optional]}

\item{\code{uint16}}{integer [optional]}

\item{\code{int32}}{integer [optional]}

\item{\code{uint32}}{integer [optional]}

\item{\code{int64}}{integer [optional]}

\item{\code{uint64}}{integer [optional]}

\item{\code{float32}}{integer [optional]}

\item{\code{float64}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DimensionTileExtent-new}{\code{DimensionTileExtent$new()}}
\item \href{#method-DimensionTileExtent-toJSON}{\code{DimensionTileExtent$toJSON()}}
\item \href{#method-DimensionTileExtent-fromJSON}{\code{DimensionTileExtent$fromJSON()}}
\item \href{#method-DimensionTileExtent-toJSONString}{\code{DimensionTileExtent$toJSONString()}}
\item \href{#method-DimensionTileExtent-fromJSONString}{\code{DimensionTileExtent$fromJSONString()}}
\item \href{#method-DimensionTileExtent-clone}{\code{DimensionTileExtent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionTileExtent-new"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionTileExtent-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionTileExtent$new(
  int8 = NULL,
  uint8 = NULL,
  int16 = NULL,
  uint16 = NULL,
  int32 = NULL,
  uint32 = NULL,
  int64 = NULL,
  uint64 = NULL,
  float32 = NULL,
  float64 = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionTileExtent-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionTileExtent-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionTileExtent$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionTileExtent-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionTileExtent-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionTileExtent$fromJSON(DimensionTileExtentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionTileExtent-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionTileExtent-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionTileExtent$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionTileExtent-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionTileExtent-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionTileExtent$fromJSONString(DimensionTileExtentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DimensionTileExtent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DimensionTileExtent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DimensionTileExtent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
