% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\docType{class}
\name{Domain}
\alias{Domain}
\title{Domain}
\format{
An \code{R6Class} generator object
}
\description{
Domain Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{\link{Datatype}}

\item{\code{tileOrder}}{\link{Layout}}

\item{\code{cellOrder}}{\link{Layout}}

\item{\code{dimensions}}{list( \link{Dimension} )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Domain-new}{\code{Domain$new()}}
\item \href{#method-Domain-toJSON}{\code{Domain$toJSON()}}
\item \href{#method-Domain-fromJSON}{\code{Domain$fromJSON()}}
\item \href{#method-Domain-toJSONString}{\code{Domain$toJSONString()}}
\item \href{#method-Domain-fromJSONString}{\code{Domain$fromJSONString()}}
\item \href{#method-Domain-clone}{\code{Domain$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Domain-new"></a>}}
\if{latex}{\out{\hypertarget{method-Domain-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Domain$new(type, tileOrder, cellOrder, dimensions, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Domain-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Domain-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Domain$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Domain-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Domain-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Domain$fromJSON(DomainJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Domain-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Domain-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Domain$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Domain-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Domain-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Domain$fromJSONString(DomainJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Domain-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Domain-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Domain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
