% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_responses.R
\name{.get_decoded_response_body_or_stop}
\alias{.get_decoded_response_body_or_stop}
\title{Package-internal HTTP-response helper}
\usage{
.get_decoded_response_body_or_stop(
  resultObject,
  result_format,
  entire_json_is_result = FALSE
)
}
\arguments{
\item{resultObject}{Should be a return value from an API function which uses
\code{\link{.wrap_as_api_response}} internally. These are functions which are manually
edited after OpenAPI autogen.}

\item{entire_json_is_result}{If false, return the \code{"value"} field from the JSON object.
This is the right thing to do for returns from the REST server for almost all cases.
The true case is only for getting the results from invoking registered Python UDFs
from R, in which case the JSON result in its entirety is the UDF output.}
}
\value{
The argument, decoded according to the specified result format.
}
\description{
This is a package-internal function for code-deduplication within various
manual-layer functions.
}
\details{
It wraps \code{\link{.get_raw_response_body_or_stop}} by decoding
the raw response body using any of the three result-format types
we support for UDFs. It's a keystroke-saving wrapper around
\code{\link{.get_raw_response_body_or_stop}}.
}
