% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_responses.R
\name{.get_empty_response_body_or_stop}
\alias{.get_empty_response_body_or_stop}
\title{Package-internal HTTP-response helper}
\usage{
.get_empty_response_body_or_stop(resultObject)
}
\arguments{
\item{resultObject}{Should be a return value from an API function which uses
\code{\link{.wrap_as_api_response}} internally. These are functions which are manually
edited after OpenAPI autogen.}
}
\value{
Invisible on success, or \code{stop()} on failure.
}
\description{
This is a package-internal function for code-deduplication within various
manual-layer functions.
}
\details{
This wraps \code{\link{.get_raw_response_body_or_stop}}, doing \code{stop}
if there is an API-response error, or if the response is not the empty string.
This is for API functions where the expected response is the empty string.
}
