% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\docType{class}
\name{Error}
\alias{Error}
\title{Error}
\format{
An \code{R6Class} generator object
}
\description{
Error Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{code}}{integer [optional]}

\item{\code{message}}{character [optional]}

\item{\code{request_id}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Error-new}{\code{Error$new()}}
\item \href{#method-Error-toJSON}{\code{Error$toJSON()}}
\item \href{#method-Error-fromJSON}{\code{Error$fromJSON()}}
\item \href{#method-Error-toJSONString}{\code{Error$toJSONString()}}
\item \href{#method-Error-fromJSONString}{\code{Error$fromJSONString()}}
\item \href{#method-Error-clone}{\code{Error$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-new"></a>}}
\if{latex}{\out{\hypertarget{method-Error-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$new(code = NULL, message = NULL, request_id = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Error-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Error-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$fromJSON(ErrorJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Error-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Error-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$fromJSONString(ErrorJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Error-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Error-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Error$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
