% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_array_udf}
\alias{execute_array_udf}
\title{TileDB Cloud UDF-Execution Helper}
\usage{
execute_array_udf(
  namespace,
  array,
  udf,
  layout = NULL,
  selectedRanges,
  attrs = NULL
)
}
\arguments{
\item{namespace}{Namespace within TileDB cloud.}

\item{array}{TODO: describe this.}

\item{udf}{An R function. TODO: more context here on that function's
signature.}

\item{queryRanges}{TODO: describe this.}
}
\value{
TODO: describe
}
\description{
This function invokes a user-defined function in TileDB Cloud.
Nominally you will first call login(); if not, the results
of the last login at ~/.tiledb/cloud.json will be used.
}
\details{
All arguments are required.
}
