% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_multi_array_udf}
\alias{execute_multi_array_udf}
\title{TileDB Cloud UDF-Execution Helper for multiple-array UDFs}
\usage{
execute_multi_array_udf(
  namespace,
  array_list,
  udf,
  layout = NULL,
  result_format = NULL
)
}
\arguments{
\item{namespace}{Namespace within TileDB cloud.}

\item{array_list}{Names of the arrays. E.g. if one URI is \code{tiledb://hello/earth}
and another is \code{tiledb://hello/mars} ##' then the namespace is "hello" and the array_list is
\code{list("earth", "mars")}.}

\item{udf}{An R function which takes an array of dataframes as argument.}

\item{queryRanges}{List of two-column matrices, one matrix per dimension,
each matrix being a start-end pair.}
}
\value{
Return value from the UDF.
}
\description{
This function invokes a user-defined function in TileDB Cloud.
Nominally you will first call login(); if not, the results
of the last login at ~/.tiledb/cloud.json will be used.
}
\details{
All arguments are required.
}
