% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_sql.R
\name{execute_sql_query}
\alias{execute_sql_query}
\title{Execute a SQL query on TileDB Cloud}
\usage{
execute_sql_query(query, name = NULL, namespace = NULL)
}
\arguments{
\item{query}{SQL query as a string.}

\item{name}{A descriptive name to give the query.}

\item{namespace}{Namespace within TileDB cloud to charge. If this is null, the
logged-in user's username will be used for the namespace.}
}
\value{
The result of the SQL query.
}
\description{
This invokes a user-defined function in TileDB Cloud.
}
\details{
Nominally you will first call \code{\link{login}}; if not, the results
of the last login at \code{~/.tiledb/cloud.json} will be used.

The \code{udf} and \code{namespace} arguments are required; the \code{args}
argument is optional.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute_sequentially}()},
\code{\link{compute}()},
\code{\link{delayed_args<-}()},
\code{\link{delayed_args}()},
\code{\link{delayed_array_udf}()},
\code{\link{delayed_generic_udf}()},
\code{\link{delayed_sql}()},
\code{\link{delayed}()},
\code{\link{deregister_array}()},
\code{\link{deregister_group}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{list_groups}()},
\code{\link{login}()},
\code{\link{register_array}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
