% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_udf}
\alias{execute_udf}
\title{TileDB Cloud Login Helper}
\usage{
execute_udf(udf, args = NULL, namespace)
}
\arguments{
\item{udf}{A R function.}

\item{args}{Arguments to the function. If the function takes
no arguments, this can be omitted. If you want to call by
position, use a list like \code{args=list(123, 456)}. If you want
to call by name, use a named list like \code{args=list(x=123,y=456)}.}

\item{namespace}{Namespace within TileDB cloud.}
}
\value{
The R object which is the return value from the UDF.
}
\description{
This function invokes a user-defined function in TileDB Cloud.
Nominally you will either call login(); if not, the results
of the last login at ~/.tiledb/cloud.json will be used.
}
\details{
The udf and namespace arguments are required; the args argument is
optional.
}
