% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_created.R
\docType{class}
\name{FileCreated}
\alias{FileCreated}
\title{FileCreated}
\format{
An \code{R6Class} generator object
}
\description{
FileCreated Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{output_uri}}{character [optional]}

\item{\code{file_name}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileCreated-new}{\code{FileCreated$new()}}
\item \href{#method-FileCreated-toJSON}{\code{FileCreated$toJSON()}}
\item \href{#method-FileCreated-fromJSON}{\code{FileCreated$fromJSON()}}
\item \href{#method-FileCreated-toJSONString}{\code{FileCreated$toJSONString()}}
\item \href{#method-FileCreated-fromJSONString}{\code{FileCreated$fromJSONString()}}
\item \href{#method-FileCreated-clone}{\code{FileCreated$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileCreated-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileCreated-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileCreated$new(output_uri = NULL, file_name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileCreated-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FileCreated-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileCreated$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileCreated-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FileCreated-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileCreated$fromJSON(FileCreatedJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileCreated-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FileCreated-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileCreated$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileCreated-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FileCreated-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileCreated$fromJSONString(FileCreatedJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileCreated-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileCreated-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileCreated$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
