% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_export.R
\docType{class}
\name{FileExport}
\alias{FileExport}
\title{FileExport}
\format{
An \code{R6Class} generator object
}
\description{
FileExport Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{output_uri}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileExport-new}{\code{FileExport$new()}}
\item \href{#method-FileExport-toJSON}{\code{FileExport$toJSON()}}
\item \href{#method-FileExport-fromJSON}{\code{FileExport$fromJSON()}}
\item \href{#method-FileExport-toJSONString}{\code{FileExport$toJSONString()}}
\item \href{#method-FileExport-fromJSONString}{\code{FileExport$fromJSONString()}}
\item \href{#method-FileExport-clone}{\code{FileExport$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileExport-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileExport-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileExport$new(output_uri = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileExport-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FileExport-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileExport$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileExport-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FileExport-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileExport$fromJSON(FileExportJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileExport-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FileExport-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileExport$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileExport-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FileExport-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileExport$fromJSONString(FileExportJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileExport-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileExport-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileExport$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
