% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_exported.R
\docType{class}
\name{FileExported}
\alias{FileExported}
\title{FileExported}
\format{
An \code{R6Class} generator object
}
\description{
FileExported Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{output_uri}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileExported-new}{\code{FileExported$new()}}
\item \href{#method-FileExported-toJSON}{\code{FileExported$toJSON()}}
\item \href{#method-FileExported-fromJSON}{\code{FileExported$fromJSON()}}
\item \href{#method-FileExported-toJSONString}{\code{FileExported$toJSONString()}}
\item \href{#method-FileExported-fromJSONString}{\code{FileExported$fromJSONString()}}
\item \href{#method-FileExported-clone}{\code{FileExported$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileExported-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileExported-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileExported$new(output_uri = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileExported-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FileExported-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileExported$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileExported-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FileExported-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileExported$fromJSON(FileExportedJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileExported-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FileExported-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileExported$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileExported-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FileExported-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileExported$fromJSONString(FileExportedJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileExported-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileExported-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileExported$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
