% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_type.R
\docType{class}
\name{FileType}
\alias{FileType}
\title{FileType}
\format{
An \code{R6Class} generator object
}
\description{
FileType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FileType-new}{\code{FileType$new()}}
\item \href{#method-FileType-toJSON}{\code{FileType$toJSON()}}
\item \href{#method-FileType-fromJSON}{\code{FileType$fromJSON()}}
\item \href{#method-FileType-toJSONString}{\code{FileType$toJSONString()}}
\item \href{#method-FileType-fromJSONString}{\code{FileType$fromJSONString()}}
\item \href{#method-FileType-clone}{\code{FileType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileType-new"></a>}}
\if{latex}{\out{\hypertarget{method-FileType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileType$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FileType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FileType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileType$fromJSON(FileTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FileType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FileType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileType$fromJSONString(FileTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FileType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FileType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FileType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
