% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\docType{class}
\name{FilterData}
\alias{FilterData}
\title{FilterData}
\format{
An \code{R6Class} generator object
}
\description{
FilterData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{int8}}{integer [optional]}

\item{\code{uint8}}{integer [optional]}

\item{\code{int16}}{integer [optional]}

\item{\code{uint16}}{integer [optional]}

\item{\code{int32}}{integer [optional]}

\item{\code{uint32}}{integer [optional]}

\item{\code{int64}}{integer [optional]}

\item{\code{uint64}}{integer [optional]}

\item{\code{float32}}{integer [optional]}

\item{\code{float64}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilterData-new}{\code{FilterData$new()}}
\item \href{#method-FilterData-toJSON}{\code{FilterData$toJSON()}}
\item \href{#method-FilterData-fromJSON}{\code{FilterData$fromJSON()}}
\item \href{#method-FilterData-toJSONString}{\code{FilterData$toJSONString()}}
\item \href{#method-FilterData-fromJSONString}{\code{FilterData$fromJSONString()}}
\item \href{#method-FilterData-clone}{\code{FilterData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterData-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilterData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterData$new(
  int8 = NULL,
  uint8 = NULL,
  int16 = NULL,
  uint16 = NULL,
  int32 = NULL,
  uint32 = NULL,
  int64 = NULL,
  uint64 = NULL,
  float32 = NULL,
  float64 = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilterData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilterData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterData$fromJSON(FilterDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilterData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilterData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterData$fromJSONString(FilterDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilterData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
