% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_option.R
\docType{class}
\name{FilterOption}
\alias{FilterOption}
\title{FilterOption}
\format{
An \code{R6Class} generator object
}
\description{
FilterOption Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilterOption-new}{\code{FilterOption$new()}}
\item \href{#method-FilterOption-toJSON}{\code{FilterOption$toJSON()}}
\item \href{#method-FilterOption-fromJSON}{\code{FilterOption$fromJSON()}}
\item \href{#method-FilterOption-toJSONString}{\code{FilterOption$toJSONString()}}
\item \href{#method-FilterOption-fromJSONString}{\code{FilterOption$fromJSONString()}}
\item \href{#method-FilterOption-clone}{\code{FilterOption$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterOption-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilterOption-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterOption$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterOption-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilterOption-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterOption$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterOption-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilterOption-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterOption$fromJSON(FilterOptionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterOption-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilterOption-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterOption$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterOption-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilterOption-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterOption$fromJSONString(FilterOptionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterOption-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilterOption-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterOption$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
