% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pipeline.R
\docType{class}
\name{FilterPipeline}
\alias{FilterPipeline}
\title{FilterPipeline}
\format{
An \code{R6Class} generator object
}
\description{
FilterPipeline Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filters}}{list( \link{Filter} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilterPipeline-new}{\code{FilterPipeline$new()}}
\item \href{#method-FilterPipeline-toJSON}{\code{FilterPipeline$toJSON()}}
\item \href{#method-FilterPipeline-fromJSON}{\code{FilterPipeline$fromJSON()}}
\item \href{#method-FilterPipeline-toJSONString}{\code{FilterPipeline$toJSONString()}}
\item \href{#method-FilterPipeline-fromJSONString}{\code{FilterPipeline$fromJSONString()}}
\item \href{#method-FilterPipeline-clone}{\code{FilterPipeline$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterPipeline-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilterPipeline-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterPipeline$new(filters = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterPipeline-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilterPipeline-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterPipeline$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterPipeline-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilterPipeline-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterPipeline$fromJSON(FilterPipelineJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterPipeline-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilterPipeline-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterPipeline$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterPipeline-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilterPipeline-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterPipeline$fromJSONString(FilterPipelineJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterPipeline-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilterPipeline-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterPipeline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
