% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_type.R
\docType{class}
\name{FilterType}
\alias{FilterType}
\title{FilterType}
\format{
An \code{R6Class} generator object
}
\description{
FilterType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FilterType-new}{\code{FilterType$new()}}
\item \href{#method-FilterType-toJSON}{\code{FilterType$toJSON()}}
\item \href{#method-FilterType-fromJSON}{\code{FilterType$fromJSON()}}
\item \href{#method-FilterType-toJSONString}{\code{FilterType$toJSONString()}}
\item \href{#method-FilterType-fromJSONString}{\code{FilterType$fromJSONString()}}
\item \href{#method-FilterType-clone}{\code{FilterType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterType-new"></a>}}
\if{latex}{\out{\hypertarget{method-FilterType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterType$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilterType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-FilterType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterType$fromJSON(FilterTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilterType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-FilterType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterType$fromJSONString(FilterTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilterType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FilterType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilterType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
