% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_udf.R
\docType{class}
\name{GenericUDF}
\alias{GenericUDF}
\title{GenericUDF}
\format{
An \code{R6Class} generator object
}
\description{
GenericUDF Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{udf_info_name}}{character [optional]}

\item{\code{language}}{\link{UDFLanguage} [optional]}

\item{\code{version}}{character [optional]}

\item{\code{image_name}}{character [optional]}

\item{\code{resource_class}}{character [optional]}

\item{\code{exec}}{character [optional]}

\item{\code{exec_raw}}{character [optional]}

\item{\code{argument}}{character [optional]}

\item{\code{stored_param_uuids}}{list( character ) [optional]}

\item{\code{result_format}}{\link{ResultFormat} [optional]}

\item{\code{task_name}}{character [optional]}

\item{\code{store_results}}{character [optional]}

\item{\code{timeout}}{integer [optional]}

\item{\code{dont_download_results}}{character [optional]}

\item{\code{task_graph_uuid}}{character [optional]}

\item{\code{client_node_uuid}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GenericUDF-new}{\code{GenericUDF$new()}}
\item \href{#method-GenericUDF-toJSON}{\code{GenericUDF$toJSON()}}
\item \href{#method-GenericUDF-fromJSON}{\code{GenericUDF$fromJSON()}}
\item \href{#method-GenericUDF-toJSONString}{\code{GenericUDF$toJSONString()}}
\item \href{#method-GenericUDF-fromJSONString}{\code{GenericUDF$fromJSONString()}}
\item \href{#method-GenericUDF-clone}{\code{GenericUDF$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericUDF-new"></a>}}
\if{latex}{\out{\hypertarget{method-GenericUDF-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericUDF$new(
  udf_info_name = NULL,
  language = NULL,
  version = NULL,
  image_name = NULL,
  resource_class = NULL,
  exec = NULL,
  exec_raw = NULL,
  argument = NULL,
  stored_param_uuids = NULL,
  result_format = NULL,
  task_name = NULL,
  store_results = NULL,
  timeout = NULL,
  dont_download_results = NULL,
  task_graph_uuid = NULL,
  client_node_uuid = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericUDF-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GenericUDF-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericUDF$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericUDF-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GenericUDF-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericUDF$fromJSON(GenericUDFJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericUDF-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GenericUDF-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericUDF$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericUDF-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GenericUDF-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericUDF$fromJSONString(GenericUDFJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenericUDF-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GenericUDF-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GenericUDF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
