% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_login.R
\name{get_api_client_instance}
\alias{get_api_client_instance}
\title{Access cached API-client object}
\usage{
get_api_client_instance()
}
\value{
The cached \code{\link{ApiClient}} object, or stops.
}
\description{
This is a package-internal function.
}
\details{
It returns the cached \code{\link{ApiClient}} object, or stops. The
\code{ApiClient} instance is constructor input to \code{\link{UserApi}},
\code{\link{ArrayApi}}, etc.  as generated by OpenAPI. Note that
with in the OpenAPI autogen code, there is the \code{ApiClient}
instance which is used as constructor input to \code{UserApi},
\code{ArrayApi}, \code{UdfAPI}, etc.
}
