% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_udf.R
\name{get_udf_info}
\alias{get_udf_info}
\title{Get information about a UDF on TileDB Cloud}
\usage{
get_udf_info(name, namespace)
}
\arguments{
\item{name}{Name of the UDF in TileDB Cloud, e.g. \code{myudfname}.}

\item{namespace}{Namespace for the UDF in TileDB Cloud, e.g. \code{mynamespace}.}
}
\value{
List of key-value pairs of UDF information.
}
\description{
Reads back information for a specified user-defined function on TileDB Cloud.
Note that \code{version}, \code{image_name}, \code{exec}, and \code{exec_raw}
are writable via \code{register_udf} but are not read back by this function.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute}()},
\code{\link{delayed}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
