% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\docType{class}
\name{Group}
\alias{Group}
\title{Group}
\format{
An \code{R6Class} generator object
}
\description{
Group Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{namespace}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{description}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Group-new}{\code{Group$new()}}
\item \href{#method-Group-toJSON}{\code{Group$toJSON()}}
\item \href{#method-Group-fromJSON}{\code{Group$fromJSON()}}
\item \href{#method-Group-toJSONString}{\code{Group$toJSONString()}}
\item \href{#method-Group-fromJSONString}{\code{Group$fromJSONString()}}
\item \href{#method-Group-clone}{\code{Group$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-new"></a>}}
\if{latex}{\out{\hypertarget{method-Group-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$new(id = NULL, namespace = NULL, name = NULL, description = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Group-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Group-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$fromJSON(GroupJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Group-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Group-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$fromJSONString(GroupJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Group-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Group-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Group$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
