% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_actions.R
\docType{class}
\name{GroupActions}
\alias{GroupActions}
\title{GroupActions}
\format{
An \code{R6Class} generator object
}
\description{
GroupActions Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupActions-new}{\code{GroupActions$new()}}
\item \href{#method-GroupActions-toJSON}{\code{GroupActions$toJSON()}}
\item \href{#method-GroupActions-fromJSON}{\code{GroupActions$fromJSON()}}
\item \href{#method-GroupActions-toJSONString}{\code{GroupActions$toJSONString()}}
\item \href{#method-GroupActions-fromJSONString}{\code{GroupActions$fromJSONString()}}
\item \href{#method-GroupActions-clone}{\code{GroupActions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupActions-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupActions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupActions$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupActions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupActions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupActions$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupActions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupActions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupActions$fromJSON(GroupActionsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupActions-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupActions-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupActions$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupActions-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupActions-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupActions$fromJSONString(GroupActionsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupActions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupActions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupActions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
