% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_browser_data.R
\docType{class}
\name{GroupBrowserData}
\alias{GroupBrowserData}
\title{GroupBrowserData}
\format{
An \code{R6Class} generator object
}
\description{
GroupBrowserData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{groups}}{list( \link{GroupInfo} ) [optional]}

\item{\code{pagination_metadata}}{\link{PaginationMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupBrowserData-new}{\code{GroupBrowserData$new()}}
\item \href{#method-GroupBrowserData-toJSON}{\code{GroupBrowserData$toJSON()}}
\item \href{#method-GroupBrowserData-fromJSON}{\code{GroupBrowserData$fromJSON()}}
\item \href{#method-GroupBrowserData-toJSONString}{\code{GroupBrowserData$toJSONString()}}
\item \href{#method-GroupBrowserData-fromJSONString}{\code{GroupBrowserData$fromJSONString()}}
\item \href{#method-GroupBrowserData-clone}{\code{GroupBrowserData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupBrowserData-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupBrowserData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupBrowserData$new(groups = NULL, pagination_metadata = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupBrowserData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupBrowserData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupBrowserData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupBrowserData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupBrowserData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupBrowserData$fromJSON(GroupBrowserDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupBrowserData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupBrowserData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupBrowserData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupBrowserData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupBrowserData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupBrowserData$fromJSONString(GroupBrowserDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupBrowserData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupBrowserData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupBrowserData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
