% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_browser_filter_data.R
\docType{class}
\name{GroupBrowserFilterData}
\alias{GroupBrowserFilterData}
\title{GroupBrowserFilterData}
\format{
An \code{R6Class} generator object
}
\description{
GroupBrowserFilterData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{namespaces}}{list( character ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupBrowserFilterData-new}{\code{GroupBrowserFilterData$new()}}
\item \href{#method-GroupBrowserFilterData-toJSON}{\code{GroupBrowserFilterData$toJSON()}}
\item \href{#method-GroupBrowserFilterData-fromJSON}{\code{GroupBrowserFilterData$fromJSON()}}
\item \href{#method-GroupBrowserFilterData-toJSONString}{\code{GroupBrowserFilterData$toJSONString()}}
\item \href{#method-GroupBrowserFilterData-fromJSONString}{\code{GroupBrowserFilterData$fromJSONString()}}
\item \href{#method-GroupBrowserFilterData-clone}{\code{GroupBrowserFilterData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupBrowserFilterData-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupBrowserFilterData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupBrowserFilterData$new(namespaces = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupBrowserFilterData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupBrowserFilterData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupBrowserFilterData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupBrowserFilterData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupBrowserFilterData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupBrowserFilterData$fromJSON(GroupBrowserFilterDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupBrowserFilterData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupBrowserFilterData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupBrowserFilterData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupBrowserFilterData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupBrowserFilterData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupBrowserFilterData$fromJSONString(GroupBrowserFilterDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupBrowserFilterData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupBrowserFilterData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupBrowserFilterData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
