% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_changes.R
\docType{class}
\name{GroupChanges}
\alias{GroupChanges}
\title{GroupChanges}
\format{
An \code{R6Class} generator object
}
\description{
GroupChanges Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{add}}{list( \link{GroupMember} ) [optional]}

\item{\code{remove}}{list( \link{GroupMember} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupChanges-new}{\code{GroupChanges$new()}}
\item \href{#method-GroupChanges-toJSON}{\code{GroupChanges$toJSON()}}
\item \href{#method-GroupChanges-fromJSON}{\code{GroupChanges$fromJSON()}}
\item \href{#method-GroupChanges-toJSONString}{\code{GroupChanges$toJSONString()}}
\item \href{#method-GroupChanges-fromJSONString}{\code{GroupChanges$fromJSONString()}}
\item \href{#method-GroupChanges-clone}{\code{GroupChanges$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupChanges-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupChanges-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupChanges$new(add = NULL, remove = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupChanges-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupChanges-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupChanges$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupChanges-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupChanges-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupChanges$fromJSON(GroupChangesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupChanges-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupChanges-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupChanges$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupChanges-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupChanges-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupChanges$fromJSONString(GroupChangesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupChanges-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupChanges-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupChanges$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
