% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_contents.R
\docType{class}
\name{GroupContents}
\alias{GroupContents}
\title{GroupContents}
\format{
An \code{R6Class} generator object
}
\description{
GroupContents Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entries}}{list( \link{GroupEntry} ) [optional]}

\item{\code{pagination_metadata}}{\link{PaginationMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupContents-new}{\code{GroupContents$new()}}
\item \href{#method-GroupContents-toJSON}{\code{GroupContents$toJSON()}}
\item \href{#method-GroupContents-fromJSON}{\code{GroupContents$fromJSON()}}
\item \href{#method-GroupContents-toJSONString}{\code{GroupContents$toJSONString()}}
\item \href{#method-GroupContents-fromJSONString}{\code{GroupContents$fromJSONString()}}
\item \href{#method-GroupContents-clone}{\code{GroupContents$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupContents-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupContents-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupContents$new(entries = NULL, pagination_metadata = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupContents-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupContents-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupContents$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupContents-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupContents-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupContents$fromJSON(GroupContentsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupContents-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupContents-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupContents$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupContents-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupContents-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupContents$fromJSONString(GroupContentsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupContents-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupContents-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupContents$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
