% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_contents_filter_data.R
\docType{class}
\name{GroupContentsFilterData}
\alias{GroupContentsFilterData}
\title{GroupContentsFilterData}
\format{
An \code{R6Class} generator object
}
\description{
GroupContentsFilterData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{namespaces}}{list( character ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupContentsFilterData-new}{\code{GroupContentsFilterData$new()}}
\item \href{#method-GroupContentsFilterData-toJSON}{\code{GroupContentsFilterData$toJSON()}}
\item \href{#method-GroupContentsFilterData-fromJSON}{\code{GroupContentsFilterData$fromJSON()}}
\item \href{#method-GroupContentsFilterData-toJSONString}{\code{GroupContentsFilterData$toJSONString()}}
\item \href{#method-GroupContentsFilterData-fromJSONString}{\code{GroupContentsFilterData$fromJSONString()}}
\item \href{#method-GroupContentsFilterData-clone}{\code{GroupContentsFilterData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupContentsFilterData-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupContentsFilterData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupContentsFilterData$new(namespaces = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupContentsFilterData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupContentsFilterData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupContentsFilterData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupContentsFilterData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupContentsFilterData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupContentsFilterData$fromJSON(GroupContentsFilterDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupContentsFilterData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupContentsFilterData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupContentsFilterData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupContentsFilterData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupContentsFilterData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupContentsFilterData$fromJSONString(GroupContentsFilterDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupContentsFilterData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupContentsFilterData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupContentsFilterData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
