% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_entry.R
\docType{class}
\name{GroupEntry}
\alias{GroupEntry}
\title{GroupEntry}
\format{
An \code{R6Class} generator object
}
\description{
GroupEntry Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{member_id}}{character [optional]}

\item{\code{group}}{\link{GroupInfo} [optional]}

\item{\code{array}}{\link{ArrayInfo} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupEntry-new}{\code{GroupEntry$new()}}
\item \href{#method-GroupEntry-toJSON}{\code{GroupEntry$toJSON()}}
\item \href{#method-GroupEntry-fromJSON}{\code{GroupEntry$fromJSON()}}
\item \href{#method-GroupEntry-toJSONString}{\code{GroupEntry$toJSONString()}}
\item \href{#method-GroupEntry-fromJSONString}{\code{GroupEntry$fromJSONString()}}
\item \href{#method-GroupEntry-clone}{\code{GroupEntry$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupEntry-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupEntry-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupEntry$new(member_id = NULL, group = NULL, array = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupEntry-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupEntry-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupEntry$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupEntry-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupEntry-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupEntry$fromJSON(GroupEntryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupEntry-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupEntry-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupEntry$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupEntry-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupEntry-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupEntry$fromJSONString(GroupEntryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupEntry-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupEntry$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
