% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_info.R
\docType{class}
\name{GroupInfo}
\alias{GroupInfo}
\title{GroupInfo}
\format{
An \code{R6Class} generator object
}
\description{
GroupInfo Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{namespace}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{uri}}{character [optional]}

\item{\code{tiledb_uri}}{character [optional]}

\item{\code{asset_count}}{numeric [optional]}

\item{\code{group_count}}{numeric [optional]}

\item{\code{size}}{numeric [optional]}

\item{\code{last_accessed}}{character [optional]}

\item{\code{allowed_actions}}{list( \link{GroupActions} ) [optional]}

\item{\code{logo}}{character [optional]}

\item{\code{access_credentials_name}}{character [optional]}

\item{\code{share_count}}{numeric [optional]}

\item{\code{public_share}}{character [optional]}

\item{\code{tags}}{list( character ) [optional]}

\item{\code{license_id}}{character [optional]}

\item{\code{license_text}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupInfo-new}{\code{GroupInfo$new()}}
\item \href{#method-GroupInfo-toJSON}{\code{GroupInfo$toJSON()}}
\item \href{#method-GroupInfo-fromJSON}{\code{GroupInfo$fromJSON()}}
\item \href{#method-GroupInfo-toJSONString}{\code{GroupInfo$toJSONString()}}
\item \href{#method-GroupInfo-fromJSONString}{\code{GroupInfo$fromJSONString()}}
\item \href{#method-GroupInfo-clone}{\code{GroupInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupInfo-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupInfo$new(
  id = NULL,
  namespace = NULL,
  name = NULL,
  description = NULL,
  uri = NULL,
  tiledb_uri = NULL,
  asset_count = NULL,
  group_count = NULL,
  size = NULL,
  last_accessed = NULL,
  allowed_actions = NULL,
  logo = NULL,
  access_credentials_name = NULL,
  share_count = NULL,
  public_share = NULL,
  tags = NULL,
  license_id = NULL,
  license_text = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupInfo$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupInfo$fromJSON(GroupInfoJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupInfo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupInfo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupInfo$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupInfo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupInfo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupInfo$fromJSONString(GroupInfoJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
