% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_listing.R
\docType{class}
\name{GroupListing}
\alias{GroupListing}
\title{GroupListing}
\format{
An \code{R6Class} generator object
}
\description{
GroupListing Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{namespace}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{description}}{character [optional]}

\item{\code{groups}}{list( \link{Group} ) [optional]}

\item{\code{assets}}{list( \link{ArrayInfo} ) [optional]}

\item{\code{pagination_metadata}}{\link{PaginationMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupListing-new}{\code{GroupListing$new()}}
\item \href{#method-GroupListing-toJSON}{\code{GroupListing$toJSON()}}
\item \href{#method-GroupListing-fromJSON}{\code{GroupListing$fromJSON()}}
\item \href{#method-GroupListing-toJSONString}{\code{GroupListing$toJSONString()}}
\item \href{#method-GroupListing-fromJSONString}{\code{GroupListing$fromJSONString()}}
\item \href{#method-GroupListing-clone}{\code{GroupListing$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupListing-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupListing-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupListing$new(
  id = NULL,
  namespace = NULL,
  name = NULL,
  description = NULL,
  groups = NULL,
  assets = NULL,
  pagination_metadata = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupListing-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupListing-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupListing$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupListing-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupListing-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupListing$fromJSON(GroupListingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupListing-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupListing-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupListing$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupListing-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupListing-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupListing$fromJSONString(GroupListingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupListing-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupListing-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupListing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
