% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_listing_all_of.R
\docType{class}
\name{GroupListingAllOf}
\alias{GroupListingAllOf}
\title{GroupListingAllOf}
\format{
An \code{R6Class} generator object
}
\description{
GroupListingAllOf Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{groups}}{list( \link{Group} ) [optional]}

\item{\code{assets}}{list( \link{ArrayInfo} ) [optional]}

\item{\code{pagination_metadata}}{\link{PaginationMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupListingAllOf-new}{\code{GroupListingAllOf$new()}}
\item \href{#method-GroupListingAllOf-toJSON}{\code{GroupListingAllOf$toJSON()}}
\item \href{#method-GroupListingAllOf-fromJSON}{\code{GroupListingAllOf$fromJSON()}}
\item \href{#method-GroupListingAllOf-toJSONString}{\code{GroupListingAllOf$toJSONString()}}
\item \href{#method-GroupListingAllOf-fromJSONString}{\code{GroupListingAllOf$fromJSONString()}}
\item \href{#method-GroupListingAllOf-clone}{\code{GroupListingAllOf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupListingAllOf-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupListingAllOf-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupListingAllOf$new(
  groups = NULL,
  assets = NULL,
  pagination_metadata = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupListingAllOf-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupListingAllOf-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupListingAllOf$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupListingAllOf-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupListingAllOf-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupListingAllOf$fromJSON(GroupListingAllOfJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupListingAllOf-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupListingAllOf-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupListingAllOf$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupListingAllOf-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupListingAllOf-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupListingAllOf$fromJSONString(GroupListingAllOfJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupListingAllOf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupListingAllOf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupListingAllOf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
