% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_member.R
\docType{class}
\name{GroupMember}
\alias{GroupMember}
\title{GroupMember}
\format{
An \code{R6Class} generator object
}
\description{
GroupMember Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{namespace}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{member_type}}{\link{GroupMemberType} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupMember-new}{\code{GroupMember$new()}}
\item \href{#method-GroupMember-toJSON}{\code{GroupMember$toJSON()}}
\item \href{#method-GroupMember-fromJSON}{\code{GroupMember$fromJSON()}}
\item \href{#method-GroupMember-toJSONString}{\code{GroupMember$toJSONString()}}
\item \href{#method-GroupMember-fromJSONString}{\code{GroupMember$fromJSONString()}}
\item \href{#method-GroupMember-clone}{\code{GroupMember$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMember-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMember-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMember$new(namespace = NULL, name = NULL, member_type = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMember-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMember-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMember$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMember-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMember-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMember$fromJSON(GroupMemberJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMember-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMember-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMember$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMember-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMember-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMember$fromJSONString(GroupMemberJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMember-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMember-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMember$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
