% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_member_asset_type.R
\docType{class}
\name{GroupMemberAssetType}
\alias{GroupMemberAssetType}
\title{GroupMemberAssetType}
\format{
An \code{R6Class} generator object
}
\description{
GroupMemberAssetType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupMemberAssetType-new}{\code{GroupMemberAssetType$new()}}
\item \href{#method-GroupMemberAssetType-toJSON}{\code{GroupMemberAssetType$toJSON()}}
\item \href{#method-GroupMemberAssetType-fromJSON}{\code{GroupMemberAssetType$fromJSON()}}
\item \href{#method-GroupMemberAssetType-toJSONString}{\code{GroupMemberAssetType$toJSONString()}}
\item \href{#method-GroupMemberAssetType-fromJSONString}{\code{GroupMemberAssetType$fromJSONString()}}
\item \href{#method-GroupMemberAssetType-clone}{\code{GroupMemberAssetType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMemberAssetType-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMemberAssetType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMemberAssetType$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMemberAssetType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMemberAssetType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMemberAssetType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMemberAssetType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMemberAssetType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMemberAssetType$fromJSON(GroupMemberAssetTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMemberAssetType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMemberAssetType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMemberAssetType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMemberAssetType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMemberAssetType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMemberAssetType$fromJSONString(GroupMemberAssetTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMemberAssetType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMemberAssetType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMemberAssetType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
