% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_member_type.R
\docType{class}
\name{GroupMemberType}
\alias{GroupMemberType}
\title{GroupMemberType}
\format{
An \code{R6Class} generator object
}
\description{
GroupMemberType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupMemberType-new}{\code{GroupMemberType$new()}}
\item \href{#method-GroupMemberType-toJSON}{\code{GroupMemberType$toJSON()}}
\item \href{#method-GroupMemberType-fromJSON}{\code{GroupMemberType$fromJSON()}}
\item \href{#method-GroupMemberType-toJSONString}{\code{GroupMemberType$toJSONString()}}
\item \href{#method-GroupMemberType-fromJSONString}{\code{GroupMemberType$fromJSONString()}}
\item \href{#method-GroupMemberType-clone}{\code{GroupMemberType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMemberType-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMemberType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMemberType$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMemberType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMemberType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMemberType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMemberType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMemberType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMemberType$fromJSON(GroupMemberTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMemberType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMemberType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMemberType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMemberType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMemberType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMemberType$fromJSONString(GroupMemberTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupMemberType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupMemberType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupMemberType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
