% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_register.R
\docType{class}
\name{GroupRegister}
\alias{GroupRegister}
\title{GroupRegister}
\format{
An \code{R6Class} generator object
}
\description{
GroupRegister Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{description}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{parent}}{character [optional]}

\item{\code{uri}}{character [optional]}

\item{\code{logo}}{character [optional]}

\item{\code{access_credentials_name}}{character [optional]}

\item{\code{tags}}{list( character ) [optional]}

\item{\code{license_id}}{character [optional]}

\item{\code{license_text}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupRegister-new}{\code{GroupRegister$new()}}
\item \href{#method-GroupRegister-toJSON}{\code{GroupRegister$toJSON()}}
\item \href{#method-GroupRegister-fromJSON}{\code{GroupRegister$fromJSON()}}
\item \href{#method-GroupRegister-toJSONString}{\code{GroupRegister$toJSONString()}}
\item \href{#method-GroupRegister-fromJSONString}{\code{GroupRegister$fromJSONString()}}
\item \href{#method-GroupRegister-clone}{\code{GroupRegister$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupRegister-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupRegister-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupRegister$new(
  description = NULL,
  name = NULL,
  parent = NULL,
  uri = NULL,
  logo = NULL,
  access_credentials_name = NULL,
  tags = NULL,
  license_id = NULL,
  license_text = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupRegister-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupRegister-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupRegister$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupRegister-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupRegister-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupRegister$fromJSON(GroupRegisterJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupRegister-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupRegister-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupRegister$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupRegister-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupRegister-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupRegister$fromJSONString(GroupRegisterJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupRegister-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupRegister-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupRegister$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
