% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_sharing.R
\docType{class}
\name{GroupSharing}
\alias{GroupSharing}
\title{GroupSharing}
\format{
An \code{R6Class} generator object
}
\description{
GroupSharing Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{group_actions}}{list( \link{GroupActions} ) [optional]}

\item{\code{array_actions}}{list( \link{ArrayActions} ) [optional]}

\item{\code{namespace}}{character [optional]}

\item{\code{namespace_type}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupSharing-new}{\code{GroupSharing$new()}}
\item \href{#method-GroupSharing-toJSON}{\code{GroupSharing$toJSON()}}
\item \href{#method-GroupSharing-fromJSON}{\code{GroupSharing$fromJSON()}}
\item \href{#method-GroupSharing-toJSONString}{\code{GroupSharing$toJSONString()}}
\item \href{#method-GroupSharing-fromJSONString}{\code{GroupSharing$fromJSONString()}}
\item \href{#method-GroupSharing-clone}{\code{GroupSharing$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupSharing-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupSharing-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupSharing$new(
  group_actions = NULL,
  array_actions = NULL,
  namespace = NULL,
  namespace_type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupSharing-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupSharing-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupSharing$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupSharing-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupSharing-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupSharing$fromJSON(GroupSharingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupSharing-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupSharing-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupSharing$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupSharing-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupSharing-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupSharing$fromJSONString(GroupSharingJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupSharing-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupSharing-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupSharing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
