% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_sharing_request.R
\docType{class}
\name{GroupSharingRequest}
\alias{GroupSharingRequest}
\title{GroupSharingRequest}
\format{
An \code{R6Class} generator object
}
\description{
GroupSharingRequest Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{group_actions}}{list( \link{GroupActions} ) [optional]}

\item{\code{array_actions}}{list( \link{ArrayActions} ) [optional]}

\item{\code{namespace}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GroupSharingRequest-new}{\code{GroupSharingRequest$new()}}
\item \href{#method-GroupSharingRequest-toJSON}{\code{GroupSharingRequest$toJSON()}}
\item \href{#method-GroupSharingRequest-fromJSON}{\code{GroupSharingRequest$fromJSON()}}
\item \href{#method-GroupSharingRequest-toJSONString}{\code{GroupSharingRequest$toJSONString()}}
\item \href{#method-GroupSharingRequest-fromJSONString}{\code{GroupSharingRequest$fromJSONString()}}
\item \href{#method-GroupSharingRequest-clone}{\code{GroupSharingRequest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupSharingRequest-new"></a>}}
\if{latex}{\out{\hypertarget{method-GroupSharingRequest-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupSharingRequest$new(
  group_actions = NULL,
  array_actions = NULL,
  namespace = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupSharingRequest-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupSharingRequest-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupSharingRequest$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupSharingRequest-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-GroupSharingRequest-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupSharingRequest$fromJSON(GroupSharingRequestJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupSharingRequest-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupSharingRequest-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupSharingRequest$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupSharingRequest-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-GroupSharingRequest-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupSharingRequest$fromJSONString(GroupSharingRequestJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GroupSharingRequest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GroupSharingRequest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GroupSharingRequest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
