% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_response200.R
\docType{class}
\name{InlineResponse200}
\alias{InlineResponse200}
\title{InlineResponse200}
\format{
An \code{R6Class} generator object
}
\description{
InlineResponse200 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{stats}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InlineResponse200-new}{\code{InlineResponse200$new()}}
\item \href{#method-InlineResponse200-toJSON}{\code{InlineResponse200$toJSON()}}
\item \href{#method-InlineResponse200-fromJSON}{\code{InlineResponse200$fromJSON()}}
\item \href{#method-InlineResponse200-toJSONString}{\code{InlineResponse200$toJSONString()}}
\item \href{#method-InlineResponse200-fromJSONString}{\code{InlineResponse200$fromJSONString()}}
\item \href{#method-InlineResponse200-clone}{\code{InlineResponse200$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-new"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$new(stats = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$fromJSON(InlineResponse200Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$fromJSONString(InlineResponse200Json)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InlineResponse200-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InlineResponse200-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InlineResponse200$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
