% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitation.R
\docType{class}
\name{Invitation}
\alias{Invitation}
\title{Invitation}
\format{
An \code{R6Class} generator object
}
\description{
Invitation Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{invitation_type}}{\link{InvitationType} [optional]}

\item{\code{owner_namespace_uuid}}{character [optional]}

\item{\code{user_namespace_uuid}}{character [optional]}

\item{\code{organization_user_uuid}}{character [optional]}

\item{\code{organization_name}}{character [optional]}

\item{\code{organization_role}}{\link{OrganizationRoles} [optional]}

\item{\code{array_uuid}}{character [optional]}

\item{\code{array_name}}{character [optional]}

\item{\code{email}}{character [optional]}

\item{\code{actions}}{character [optional]}

\item{\code{status}}{\link{InvitationStatus} [optional]}

\item{\code{created_at}}{character [optional]}

\item{\code{expires_at}}{character [optional]}

\item{\code{accepted_at}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Invitation-new}{\code{Invitation$new()}}
\item \href{#method-Invitation-toJSON}{\code{Invitation$toJSON()}}
\item \href{#method-Invitation-fromJSON}{\code{Invitation$fromJSON()}}
\item \href{#method-Invitation-toJSONString}{\code{Invitation$toJSONString()}}
\item \href{#method-Invitation-fromJSONString}{\code{Invitation$fromJSONString()}}
\item \href{#method-Invitation-clone}{\code{Invitation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Invitation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Invitation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Invitation$new(
  id = NULL,
  invitation_type = NULL,
  owner_namespace_uuid = NULL,
  user_namespace_uuid = NULL,
  organization_user_uuid = NULL,
  organization_name = NULL,
  organization_role = NULL,
  array_uuid = NULL,
  array_name = NULL,
  email = NULL,
  actions = NULL,
  status = NULL,
  created_at = NULL,
  expires_at = NULL,
  accepted_at = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Invitation-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Invitation-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Invitation$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Invitation-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Invitation-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Invitation$fromJSON(InvitationJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Invitation-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Invitation-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Invitation$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Invitation-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Invitation-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Invitation$fromJSONString(InvitationJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Invitation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Invitation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Invitation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
