% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitation_array_share_email.R
\docType{class}
\name{InvitationArrayShareEmail}
\alias{InvitationArrayShareEmail}
\title{InvitationArrayShareEmail}
\format{
An \code{R6Class} generator object
}
\description{
InvitationArrayShareEmail Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actions}}{list( \link{ArrayActions} )}

\item{\code{invitee_email}}{list( character )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InvitationArrayShareEmail-new}{\code{InvitationArrayShareEmail$new()}}
\item \href{#method-InvitationArrayShareEmail-toJSON}{\code{InvitationArrayShareEmail$toJSON()}}
\item \href{#method-InvitationArrayShareEmail-fromJSON}{\code{InvitationArrayShareEmail$fromJSON()}}
\item \href{#method-InvitationArrayShareEmail-toJSONString}{\code{InvitationArrayShareEmail$toJSONString()}}
\item \href{#method-InvitationArrayShareEmail-fromJSONString}{\code{InvitationArrayShareEmail$fromJSONString()}}
\item \href{#method-InvitationArrayShareEmail-clone}{\code{InvitationArrayShareEmail$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationArrayShareEmail-new"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationArrayShareEmail-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationArrayShareEmail$new(actions, invitee_email, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationArrayShareEmail-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationArrayShareEmail-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationArrayShareEmail$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationArrayShareEmail-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationArrayShareEmail-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationArrayShareEmail$fromJSON(InvitationArrayShareEmailJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationArrayShareEmail-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationArrayShareEmail-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationArrayShareEmail$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationArrayShareEmail-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationArrayShareEmail-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationArrayShareEmail$fromJSONString(InvitationArrayShareEmailJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationArrayShareEmail-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationArrayShareEmail-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationArrayShareEmail$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
