% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitation_organization_join_email.R
\docType{class}
\name{InvitationOrganizationJoinEmail}
\alias{InvitationOrganizationJoinEmail}
\title{InvitationOrganizationJoinEmail}
\format{
An \code{R6Class} generator object
}
\description{
InvitationOrganizationJoinEmail Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{actions}}{list( \link{NamespaceActions} ) [optional]}

\item{\code{organization_role}}{\link{OrganizationRoles}}

\item{\code{invitee_email}}{list( character )}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InvitationOrganizationJoinEmail-new}{\code{InvitationOrganizationJoinEmail$new()}}
\item \href{#method-InvitationOrganizationJoinEmail-toJSON}{\code{InvitationOrganizationJoinEmail$toJSON()}}
\item \href{#method-InvitationOrganizationJoinEmail-fromJSON}{\code{InvitationOrganizationJoinEmail$fromJSON()}}
\item \href{#method-InvitationOrganizationJoinEmail-toJSONString}{\code{InvitationOrganizationJoinEmail$toJSONString()}}
\item \href{#method-InvitationOrganizationJoinEmail-fromJSONString}{\code{InvitationOrganizationJoinEmail$fromJSONString()}}
\item \href{#method-InvitationOrganizationJoinEmail-clone}{\code{InvitationOrganizationJoinEmail$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationOrganizationJoinEmail-new"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationOrganizationJoinEmail-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationOrganizationJoinEmail$new(
  organization_role,
  invitee_email,
  actions = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationOrganizationJoinEmail-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationOrganizationJoinEmail-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationOrganizationJoinEmail$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationOrganizationJoinEmail-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationOrganizationJoinEmail-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationOrganizationJoinEmail$fromJSON(InvitationOrganizationJoinEmailJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationOrganizationJoinEmail-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationOrganizationJoinEmail-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationOrganizationJoinEmail$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationOrganizationJoinEmail-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationOrganizationJoinEmail-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationOrganizationJoinEmail$fromJSONString(
  InvitationOrganizationJoinEmailJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationOrganizationJoinEmail-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationOrganizationJoinEmail-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationOrganizationJoinEmail$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
