% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invitation_type.R
\docType{class}
\name{InvitationType}
\alias{InvitationType}
\title{InvitationType}
\format{
An \code{R6Class} generator object
}
\description{
InvitationType Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InvitationType-new}{\code{InvitationType$new()}}
\item \href{#method-InvitationType-toJSON}{\code{InvitationType$toJSON()}}
\item \href{#method-InvitationType-fromJSON}{\code{InvitationType$fromJSON()}}
\item \href{#method-InvitationType-toJSONString}{\code{InvitationType$toJSONString()}}
\item \href{#method-InvitationType-fromJSONString}{\code{InvitationType$fromJSONString()}}
\item \href{#method-InvitationType-clone}{\code{InvitationType$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationType-new"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationType$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationType-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationType-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationType$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationType-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationType-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationType$fromJSON(InvitationTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationType-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationType-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationType$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationType-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationType-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationType$fromJSONString(InvitationTypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InvitationType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InvitationType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InvitationType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
