% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_accessed_array.R
\docType{class}
\name{LastAccessedArray}
\alias{LastAccessedArray}
\title{LastAccessedArray}
\format{
An \code{R6Class} generator object
}
\description{
LastAccessedArray Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{array_id}}{character [optional]}

\item{\code{array_name}}{character [optional]}

\item{\code{namespace}}{character [optional]}

\item{\code{accessed_time}}{numeric [optional]}

\item{\code{access_type}}{\link{ActivityEventType} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LastAccessedArray-new}{\code{LastAccessedArray$new()}}
\item \href{#method-LastAccessedArray-toJSON}{\code{LastAccessedArray$toJSON()}}
\item \href{#method-LastAccessedArray-fromJSON}{\code{LastAccessedArray$fromJSON()}}
\item \href{#method-LastAccessedArray-toJSONString}{\code{LastAccessedArray$toJSONString()}}
\item \href{#method-LastAccessedArray-fromJSONString}{\code{LastAccessedArray$fromJSONString()}}
\item \href{#method-LastAccessedArray-clone}{\code{LastAccessedArray$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LastAccessedArray-new"></a>}}
\if{latex}{\out{\hypertarget{method-LastAccessedArray-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LastAccessedArray$new(
  array_id = NULL,
  array_name = NULL,
  namespace = NULL,
  accessed_time = NULL,
  access_type = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LastAccessedArray-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LastAccessedArray-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LastAccessedArray$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LastAccessedArray-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-LastAccessedArray-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LastAccessedArray$fromJSON(LastAccessedArrayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LastAccessedArray-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LastAccessedArray-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LastAccessedArray$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LastAccessedArray-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-LastAccessedArray-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LastAccessedArray$fromJSONString(LastAccessedArrayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LastAccessedArray-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LastAccessedArray-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LastAccessedArray$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
