% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\docType{class}
\name{Layout}
\alias{Layout}
\title{Layout}
\format{
An \code{R6Class} generator object
}
\description{
Layout Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Layout-new}{\code{Layout$new()}}
\item \href{#method-Layout-toJSON}{\code{Layout$toJSON()}}
\item \href{#method-Layout-fromJSON}{\code{Layout$fromJSON()}}
\item \href{#method-Layout-toJSONString}{\code{Layout$toJSONString()}}
\item \href{#method-Layout-fromJSONString}{\code{Layout$fromJSONString()}}
\item \href{#method-Layout-clone}{\code{Layout$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layout-new"></a>}}
\if{latex}{\out{\hypertarget{method-Layout-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layout$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layout-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Layout-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layout$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layout-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Layout-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layout$fromJSON(LayoutJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layout-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Layout-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layout$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layout-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Layout-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layout$fromJSONString(LayoutJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Layout-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Layout-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Layout$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
