% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_array.R
\name{list_arrays}
\alias{list_arrays}
\title{TODO: comment
xxx note this is a paginable API but default params are all-at-once}
\usage{
list_arrays(namespace = NULL, permissions = NULL)
}
\description{
TODO: comment
xxx note this is a paginable API but default params are all-at-once
}
