% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_group.R
\name{list_groups}
\alias{list_groups}
\title{Show listing of groups}
\usage{
list_groups(
  public = FALSE,
  shared = FALSE,
  page = NULL,
  per.page = NULL,
  search = NULL,
  namespace = NULL,
  orderby = NULL,
  permissions = NULL,
  tag = NULL,
  exclude.tag = NULL,
  flat = FALSE,
  parent = NULL
)
}
\arguments{
\item{public}{logical TRUE means list public arrays}

\item{shared}{logical TRUE means list shared arrays. If \code{public} and \code{shared} are both
false then arrays owned by you are listed.}

\item{page}{integer}

\item{per.page}{integer}

\item{search}{character}

\item{namespace}{character}

\item{orderby}{character}

\item{permissions}{character}

\item{tag}{list( character )}

\item{exclude.tag}{list( character )}

\item{flat}{logical, if `TRUE`, ignores the nesting of groups and searches
all of them}

\item{parent}{character, search only the children of the groups with this
uuid}
}
\value{
A `data.frame` of metadata for all groups in your account that meet
the filter applied.
}
\description{
Returns a dataframe of metadata for all groups that meet the
filter applied.
}
\details{
Note that this is a paginable API but default params return all results on
one call, even hundreds of them.

Pagination information is set as an attribute of the returned data frame.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute_sequentially}()},
\code{\link{compute}()},
\code{\link{delayed_args<-}()},
\code{\link{delayed_args}()},
\code{\link{delayed_array_udf}()},
\code{\link{delayed_generic_udf}()},
\code{\link{delayed_sql}()},
\code{\link{delayed}()},
\code{\link{deregister_array}()},
\code{\link{deregister_group}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{group_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_array}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
