% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{login}
\alias{login}
\title{Set up TileDB Cloud Login Information}
\usage{
login(
  username = "",
  password = "",
  token = "",
  remember_me = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{username}{A character value with the username, if present password is also needed.}

\item{password}{A character value with the password, if present username is also needed.}

\item{token}{A character value with the access token, it can be used instead of username
and password.}

\item{remember_me}{A boolean to select a session with for 24 hours instead of 8 hours, used
only when a new session is requested.}

\item{verbose}{A boolean switch to select whether an warning message is printed or not.}
}
\value{
Nothing is returned, the function is called for a side effect of storing the
values in the package environment
}
\description{
TileDB Cloud Login Helper
}
\details{
This function can be used to override the default setup made at package load.
It can operate in two modes. Either, a username and a password are supplied as
environment variable \code{TILEDB_REST_USERNAME} and \code{TILEDB_REST_PASSWORD}.
As an alternative, an access token can be supplied via \code{TILEDB_REST_TOKEN}.
The values are used to instantiate a new API client object. If no token was supplied
a new session is requested and the token assigned to that session is used.
}
