% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model_favorites_data.R
\docType{class}
\name{MLModelFavoritesData}
\alias{MLModelFavoritesData}
\title{MLModelFavoritesData}
\format{
An \code{R6Class} generator object
}
\description{
MLModelFavoritesData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mlmodels}}{list( \link{ArrayInfo} ) [optional]}

\item{\code{pagination_metadata}}{\link{PaginationMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLModelFavoritesData-new}{\code{MLModelFavoritesData$new()}}
\item \href{#method-MLModelFavoritesData-toJSON}{\code{MLModelFavoritesData$toJSON()}}
\item \href{#method-MLModelFavoritesData-fromJSON}{\code{MLModelFavoritesData$fromJSON()}}
\item \href{#method-MLModelFavoritesData-toJSONString}{\code{MLModelFavoritesData$toJSONString()}}
\item \href{#method-MLModelFavoritesData-fromJSONString}{\code{MLModelFavoritesData$fromJSONString()}}
\item \href{#method-MLModelFavoritesData-clone}{\code{MLModelFavoritesData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModelFavoritesData-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLModelFavoritesData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModelFavoritesData$new(mlmodels = NULL, pagination_metadata = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModelFavoritesData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MLModelFavoritesData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModelFavoritesData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModelFavoritesData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MLModelFavoritesData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModelFavoritesData$fromJSON(MLModelFavoritesDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModelFavoritesData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MLModelFavoritesData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModelFavoritesData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModelFavoritesData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MLModelFavoritesData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModelFavoritesData$fromJSONString(MLModelFavoritesDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModelFavoritesData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLModelFavoritesData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModelFavoritesData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
