% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_array_udf.R
\docType{class}
\name{MultiArrayUDF}
\alias{MultiArrayUDF}
\title{MultiArrayUDF}
\format{
An \code{R6Class} generator object
}
\description{
MultiArrayUDF Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{udf_info_name}}{character [optional]}

\item{\code{language}}{\link{UDFLanguage} [optional]}

\item{\code{version}}{character [optional]}

\item{\code{image_name}}{character [optional]}

\item{\code{exec}}{character [optional]}

\item{\code{exec_raw}}{character [optional]}

\item{\code{result_format}}{\link{ResultFormat} [optional]}

\item{\code{task_name}}{character [optional]}

\item{\code{argument}}{character [optional]}

\item{\code{stored_param_uuids}}{list( character ) [optional]}

\item{\code{store_results}}{character [optional]}

\item{\code{dont_download_results}}{character [optional]}

\item{\code{ranges}}{\link{QueryRanges} [optional]}

\item{\code{subarray}}{\link{UDFSubarray} [optional]}

\item{\code{buffers}}{list( character ) [optional]}

\item{\code{arrays}}{list( \link{UDFArrayDetails} ) [optional]}

\item{\code{timeout}}{integer [optional]}

\item{\code{task_graph_uuid}}{character [optional]}

\item{\code{client_node_uuid}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MultiArrayUDF$new()}}
\item \href{#method-toJSON}{\code{MultiArrayUDF$toJSON()}}
\item \href{#method-fromJSON}{\code{MultiArrayUDF$fromJSON()}}
\item \href{#method-toJSONString}{\code{MultiArrayUDF$toJSONString()}}
\item \href{#method-fromJSONString}{\code{MultiArrayUDF$fromJSONString()}}
\item \href{#method-clone}{\code{MultiArrayUDF$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiArrayUDF$new(
  udf_info_name = NULL,
  language = NULL,
  version = NULL,
  image_name = NULL,
  exec = NULL,
  exec_raw = NULL,
  result_format = NULL,
  task_name = NULL,
  argument = NULL,
  stored_param_uuids = NULL,
  store_results = NULL,
  dont_download_results = NULL,
  ranges = NULL,
  subarray = NULL,
  buffers = NULL,
  arrays = NULL,
  timeout = NULL,
  task_graph_uuid = NULL,
  client_node_uuid = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiArrayUDF$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiArrayUDF$fromJSON(MultiArrayUDFJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiArrayUDF$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiArrayUDF$fromJSONString(MultiArrayUDFJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiArrayUDF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
