% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace_actions.R
\docType{class}
\name{NamespaceActions}
\alias{NamespaceActions}
\title{NamespaceActions}
\format{
An \code{R6Class} generator object
}
\description{
NamespaceActions Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NamespaceActions-new}{\code{NamespaceActions$new()}}
\item \href{#method-NamespaceActions-toJSON}{\code{NamespaceActions$toJSON()}}
\item \href{#method-NamespaceActions-fromJSON}{\code{NamespaceActions$fromJSON()}}
\item \href{#method-NamespaceActions-toJSONString}{\code{NamespaceActions$toJSONString()}}
\item \href{#method-NamespaceActions-fromJSONString}{\code{NamespaceActions$fromJSONString()}}
\item \href{#method-NamespaceActions-clone}{\code{NamespaceActions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NamespaceActions-new"></a>}}
\if{latex}{\out{\hypertarget{method-NamespaceActions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamespaceActions$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NamespaceActions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NamespaceActions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamespaceActions$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NamespaceActions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NamespaceActions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamespaceActions$fromJSON(NamespaceActionsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NamespaceActions-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NamespaceActions-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamespaceActions$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NamespaceActions-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NamespaceActions-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamespaceActions$fromJSONString(NamespaceActionsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NamespaceActions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NamespaceActions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NamespaceActions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
