% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_empty_domain.R
\docType{class}
\name{NonEmptyDomain}
\alias{NonEmptyDomain}
\title{NonEmptyDomain}
\format{
An \code{R6Class} generator object
}
\description{
NonEmptyDomain Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nonEmptyDomain}}{\link{DomainArray}}

\item{\code{isEmpty}}{character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NonEmptyDomain-new}{\code{NonEmptyDomain$new()}}
\item \href{#method-NonEmptyDomain-toJSON}{\code{NonEmptyDomain$toJSON()}}
\item \href{#method-NonEmptyDomain-fromJSON}{\code{NonEmptyDomain$fromJSON()}}
\item \href{#method-NonEmptyDomain-toJSONString}{\code{NonEmptyDomain$toJSONString()}}
\item \href{#method-NonEmptyDomain-fromJSONString}{\code{NonEmptyDomain$fromJSONString()}}
\item \href{#method-NonEmptyDomain-clone}{\code{NonEmptyDomain$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonEmptyDomain-new"></a>}}
\if{latex}{\out{\hypertarget{method-NonEmptyDomain-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonEmptyDomain$new(nonEmptyDomain, isEmpty, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonEmptyDomain-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NonEmptyDomain-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonEmptyDomain$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonEmptyDomain-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NonEmptyDomain-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonEmptyDomain$fromJSON(NonEmptyDomainJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonEmptyDomain-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NonEmptyDomain-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonEmptyDomain$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonEmptyDomain-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NonEmptyDomain-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonEmptyDomain$fromJSONString(NonEmptyDomainJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NonEmptyDomain-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NonEmptyDomain-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NonEmptyDomain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
