% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook_api.R
\docType{class}
\name{NotebookApi}
\alias{NotebookApi}
\title{Notebook operations}
\format{
An \code{R6Class} generator object
}
\description{
tiledbcloud.Notebook
}
\section{Methods}{

\describe{
\strong{ GetNotebookServerStatus } \emph{  }
Get status of the notebook server

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @returnType } \link{NotebookStatus} \cr


\item status code : 200 | status of running notebook

\item return type : NotebookStatus 
\item response headers :

\tabular{ll}{
}
\item status code : 202 | Notebook server is pending


\item response headers :

\tabular{ll}{
}
\item status code : 402 | Payment required

\item return type : Error 
\item response headers :

\tabular{ll}{
}
\item status code : 404 | Notebook is not running


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ShutdownNotebookServer } \emph{  }
Shutdown a notebook server

\itemize{
\item \emph{ @param } namespace character


\item status code : 204 | Notebook shutdown successfully


\item response headers :

\tabular{ll}{
}
\item status code : 404 | Notebook is not running


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ UpdateNotebookName } \emph{  }
update name on a notebok, moving related S3 object to new location

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } notebook.metadata \link{ArrayInfoUpdate}


\item status code : 204 | notebook name updated successfully


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  GetNotebookServerStatus  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace notebook is in (an organization name or user's username)

api.instance <- NotebookApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetNotebookServerStatus(var.namespace)


####################  ShutdownNotebookServer  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace notebook is in (an organization name or user's username)

api.instance <- NotebookApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ShutdownNotebookServer(var.namespace)


####################  UpdateNotebookName  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of notebook (array) that is url-encoded
var.notebook.metadata <- ArrayInfoUpdate$new() # ArrayInfoUpdate | notebook (array) metadata to update

api.instance <- NotebookApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$UpdateNotebookName(var.namespace, var.array, var.notebook.metadata)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotebookApi-new}{\code{NotebookApi$new()}}
\item \href{#method-NotebookApi-GetNotebookServerStatus}{\code{NotebookApi$GetNotebookServerStatus()}}
\item \href{#method-NotebookApi-GetNotebookServerStatusWithHttpInfo}{\code{NotebookApi$GetNotebookServerStatusWithHttpInfo()}}
\item \href{#method-NotebookApi-ShutdownNotebookServer}{\code{NotebookApi$ShutdownNotebookServer()}}
\item \href{#method-NotebookApi-ShutdownNotebookServerWithHttpInfo}{\code{NotebookApi$ShutdownNotebookServerWithHttpInfo()}}
\item \href{#method-NotebookApi-UpdateNotebookName}{\code{NotebookApi$UpdateNotebookName()}}
\item \href{#method-NotebookApi-UpdateNotebookNameWithHttpInfo}{\code{NotebookApi$UpdateNotebookNameWithHttpInfo()}}
\item \href{#method-NotebookApi-clone}{\code{NotebookApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookApi-GetNotebookServerStatus"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookApi-GetNotebookServerStatus}{}}}
\subsection{Method \code{GetNotebookServerStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookApi$GetNotebookServerStatus(namespace, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookApi-GetNotebookServerStatusWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookApi-GetNotebookServerStatusWithHttpInfo}{}}}
\subsection{Method \code{GetNotebookServerStatusWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookApi$GetNotebookServerStatusWithHttpInfo(namespace, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookApi-ShutdownNotebookServer"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookApi-ShutdownNotebookServer}{}}}
\subsection{Method \code{ShutdownNotebookServer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookApi$ShutdownNotebookServer(namespace, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookApi-ShutdownNotebookServerWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookApi-ShutdownNotebookServerWithHttpInfo}{}}}
\subsection{Method \code{ShutdownNotebookServerWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookApi$ShutdownNotebookServerWithHttpInfo(namespace, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookApi-UpdateNotebookName"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookApi-UpdateNotebookName}{}}}
\subsection{Method \code{UpdateNotebookName()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookApi$UpdateNotebookName(namespace, array, notebook.metadata, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookApi-UpdateNotebookNameWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookApi-UpdateNotebookNameWithHttpInfo}{}}}
\subsection{Method \code{UpdateNotebookNameWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookApi$UpdateNotebookNameWithHttpInfo(
  namespace,
  array,
  notebook.metadata,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
