% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook_favorite.R
\docType{class}
\name{NotebookFavorite}
\alias{NotebookFavorite}
\title{NotebookFavorite}
\format{
An \code{R6Class} generator object
}
\description{
NotebookFavorite Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{notebook_uuid}}{character [optional]}

\item{\code{namespace}}{character [optional]}

\item{\code{name}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotebookFavorite-new}{\code{NotebookFavorite$new()}}
\item \href{#method-NotebookFavorite-toJSON}{\code{NotebookFavorite$toJSON()}}
\item \href{#method-NotebookFavorite-fromJSON}{\code{NotebookFavorite$fromJSON()}}
\item \href{#method-NotebookFavorite-toJSONString}{\code{NotebookFavorite$toJSONString()}}
\item \href{#method-NotebookFavorite-fromJSONString}{\code{NotebookFavorite$fromJSONString()}}
\item \href{#method-NotebookFavorite-clone}{\code{NotebookFavorite$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookFavorite-new"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookFavorite-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookFavorite$new(notebook_uuid = NULL, namespace = NULL, name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookFavorite-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookFavorite-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookFavorite$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookFavorite-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookFavorite-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookFavorite$fromJSON(NotebookFavoriteJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookFavorite-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookFavorite-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookFavorite$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookFavorite-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookFavorite-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookFavorite$fromJSONString(NotebookFavoriteJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookFavorite-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookFavorite-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookFavorite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
