% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebooks_api.R
\docType{class}
\name{NotebooksApi}
\alias{NotebooksApi}
\title{Notebooks operations}
\format{
An \code{R6Class} generator object
}
\description{
tiledbcloud.Notebooks
}
\section{Methods}{

\describe{
\strong{ NotebooksNamespaceArrayEndTimestampsGet } \emph{  }
retrieve a list of timestamps from the array fragment info listing in milliseconds, paginated

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } array character
\item \emph{ @param } page integer
\item \emph{ @param } per.page integer
\item \emph{ @returnType } \link{ArrayEndTimestampData} \cr


\item status code : 200 | list of timestamps in milliseconds, paginated

\item return type : ArrayEndTimestampData 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  NotebooksNamespaceArrayEndTimestampsGet  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | namespace array is in (an organization name or user's username)
var.array <- 'array_example' # character | name/uri of array that is url-encoded
var.page <- 56 # integer | pagination offset
var.per.page <- 56 # integer | pagination limit

api.instance <- NotebooksApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$NotebooksNamespaceArrayEndTimestampsGet(var.namespace, var.array, page=var.page, per.page=var.per.page)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NotebooksApi$new()}}
\item \href{#method-NotebooksNamespaceArrayEndTimestampsGet}{\code{NotebooksApi$NotebooksNamespaceArrayEndTimestampsGet()}}
\item \href{#method-NotebooksNamespaceArrayEndTimestampsGetWithHttpInfo}{\code{NotebooksApi$NotebooksNamespaceArrayEndTimestampsGetWithHttpInfo()}}
\item \href{#method-clone}{\code{NotebooksApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebooksApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebooksNamespaceArrayEndTimestampsGet"></a>}}
\if{latex}{\out{\hypertarget{method-NotebooksNamespaceArrayEndTimestampsGet}{}}}
\subsection{Method \code{NotebooksNamespaceArrayEndTimestampsGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebooksApi$NotebooksNamespaceArrayEndTimestampsGet(
  namespace,
  array,
  page = NULL,
  per.page = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebooksNamespaceArrayEndTimestampsGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-NotebooksNamespaceArrayEndTimestampsGetWithHttpInfo}{}}}
\subsection{Method \code{NotebooksNamespaceArrayEndTimestampsGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebooksApi$NotebooksNamespaceArrayEndTimestampsGetWithHttpInfo(
  namespace,
  array,
  page = NULL,
  per.page = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebooksApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
