% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook_status.R
\docType{class}
\name{NotebookStatus}
\alias{NotebookStatus}
\title{NotebookStatus}
\format{
An \code{R6Class} generator object
}
\description{
NotebookStatus Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{namespace}}{character [optional]}

\item{\code{uptime}}{integer [optional]}

\item{\code{cpu_usage}}{integer [optional]}

\item{\code{memory_usage}}{integer [optional]}

\item{\code{memory_limit}}{integer [optional]}

\item{\code{cpu_count}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NotebookStatus-new}{\code{NotebookStatus$new()}}
\item \href{#method-NotebookStatus-toJSON}{\code{NotebookStatus$toJSON()}}
\item \href{#method-NotebookStatus-fromJSON}{\code{NotebookStatus$fromJSON()}}
\item \href{#method-NotebookStatus-toJSONString}{\code{NotebookStatus$toJSONString()}}
\item \href{#method-NotebookStatus-fromJSONString}{\code{NotebookStatus$fromJSONString()}}
\item \href{#method-NotebookStatus-clone}{\code{NotebookStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookStatus-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookStatus$new(
  namespace = NULL,
  uptime = NULL,
  cpu_usage = NULL,
  memory_usage = NULL,
  memory_limit = NULL,
  cpu_count = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookStatus$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookStatus$fromJSON(NotebookStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookStatus$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookStatus$fromJSONString(NotebookStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NotebookStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NotebookStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NotebookStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
