% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricing_currency.R
\docType{class}
\name{PricingCurrency}
\alias{PricingCurrency}
\title{PricingCurrency}
\format{
An \code{R6Class} generator object
}
\description{
PricingCurrency Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PricingCurrency-new}{\code{PricingCurrency$new()}}
\item \href{#method-PricingCurrency-toJSON}{\code{PricingCurrency$toJSON()}}
\item \href{#method-PricingCurrency-fromJSON}{\code{PricingCurrency$fromJSON()}}
\item \href{#method-PricingCurrency-toJSONString}{\code{PricingCurrency$toJSONString()}}
\item \href{#method-PricingCurrency-fromJSONString}{\code{PricingCurrency$fromJSONString()}}
\item \href{#method-PricingCurrency-clone}{\code{PricingCurrency$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PricingCurrency-new"></a>}}
\if{latex}{\out{\hypertarget{method-PricingCurrency-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PricingCurrency$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PricingCurrency-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PricingCurrency-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PricingCurrency$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PricingCurrency-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PricingCurrency-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PricingCurrency$fromJSON(PricingCurrencyJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PricingCurrency-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PricingCurrency-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PricingCurrency$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PricingCurrency-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PricingCurrency-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PricingCurrency$fromJSONString(PricingCurrencyJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PricingCurrency-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PricingCurrency-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PricingCurrency$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
