% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_share_filter.R
\docType{class}
\name{PublicShareFilter}
\alias{PublicShareFilter}
\title{PublicShareFilter}
\format{
An \code{R6Class} generator object
}
\description{
PublicShareFilter Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PublicShareFilter-new}{\code{PublicShareFilter$new()}}
\item \href{#method-PublicShareFilter-toJSON}{\code{PublicShareFilter$toJSON()}}
\item \href{#method-PublicShareFilter-fromJSON}{\code{PublicShareFilter$fromJSON()}}
\item \href{#method-PublicShareFilter-toJSONString}{\code{PublicShareFilter$toJSONString()}}
\item \href{#method-PublicShareFilter-fromJSONString}{\code{PublicShareFilter$fromJSONString()}}
\item \href{#method-PublicShareFilter-clone}{\code{PublicShareFilter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PublicShareFilter-new"></a>}}
\if{latex}{\out{\hypertarget{method-PublicShareFilter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PublicShareFilter$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PublicShareFilter-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PublicShareFilter-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PublicShareFilter$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PublicShareFilter-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-PublicShareFilter-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PublicShareFilter$fromJSON(PublicShareFilterJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PublicShareFilter-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PublicShareFilter-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PublicShareFilter$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PublicShareFilter-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-PublicShareFilter-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PublicShareFilter$fromJSONString(PublicShareFilterJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PublicShareFilter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PublicShareFilter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PublicShareFilter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
