% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\docType{class}
\name{Query}
\alias{Query}
\title{Query}
\format{
An \code{R6Class} generator object
}
\description{
Query Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{\link{Querytype}}

\item{\code{layout}}{\link{Layout}}

\item{\code{status}}{\link{Querystatus}}

\item{\code{attributeBufferHeaders}}{list( \link{AttributeBufferHeader} )}

\item{\code{writer}}{\link{Writer} [optional]}

\item{\code{reader}}{\link{QueryReader} [optional]}

\item{\code{array}}{\link{Array}}

\item{\code{totalFixedLengthBufferBytes}}{integer}

\item{\code{totalVarLenBufferBytes}}{integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Query-new}{\code{Query$new()}}
\item \href{#method-Query-toJSON}{\code{Query$toJSON()}}
\item \href{#method-Query-fromJSON}{\code{Query$fromJSON()}}
\item \href{#method-Query-toJSONString}{\code{Query$toJSONString()}}
\item \href{#method-Query-fromJSONString}{\code{Query$fromJSONString()}}
\item \href{#method-Query-clone}{\code{Query$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-new"></a>}}
\if{latex}{\out{\hypertarget{method-Query-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$new(
  type,
  layout,
  status,
  attributeBufferHeaders,
  array,
  totalFixedLengthBufferBytes,
  totalVarLenBufferBytes,
  writer = NULL,
  reader = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Query-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Query-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$fromJSON(QueryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Query-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Query-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$fromJSONString(QueryJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Query-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Query-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Query$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
