% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_json.R
\docType{class}
\name{QueryJson}
\alias{QueryJson}
\title{QueryJson}
\format{
An \code{R6Class} generator object
}
\description{
QueryJson Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{query_ranges}}{\link{QueryRanges} [optional]}

\item{\code{fields}}{list( character ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QueryJson-new}{\code{QueryJson$new()}}
\item \href{#method-QueryJson-toJSON}{\code{QueryJson$toJSON()}}
\item \href{#method-QueryJson-fromJSON}{\code{QueryJson$fromJSON()}}
\item \href{#method-QueryJson-toJSONString}{\code{QueryJson$toJSONString()}}
\item \href{#method-QueryJson-fromJSONString}{\code{QueryJson$fromJSONString()}}
\item \href{#method-QueryJson-clone}{\code{QueryJson$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryJson-new"></a>}}
\if{latex}{\out{\hypertarget{method-QueryJson-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryJson$new(query_ranges = NULL, fields = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryJson-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QueryJson-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryJson$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryJson-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QueryJson-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryJson$fromJSON(QueryJsonJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryJson-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QueryJson-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryJson$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryJson-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QueryJson-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryJson$fromJSONString(QueryJsonJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryJson-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QueryJson-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryJson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
