% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ranges.R
\docType{class}
\name{QueryRanges}
\alias{QueryRanges}
\title{QueryRanges}
\format{
An \code{R6Class} generator object
}
\description{
QueryRanges Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{layout}}{\link{Layout} [optional]}

\item{\code{ranges}}{list( \link{array[numeric]} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QueryRanges-new}{\code{QueryRanges$new()}}
\item \href{#method-QueryRanges-toJSON}{\code{QueryRanges$toJSON()}}
\item \href{#method-QueryRanges-fromJSON}{\code{QueryRanges$fromJSON()}}
\item \href{#method-QueryRanges-toJSONString}{\code{QueryRanges$toJSONString()}}
\item \href{#method-QueryRanges-fromJSONString}{\code{QueryRanges$fromJSONString()}}
\item \href{#method-QueryRanges-clone}{\code{QueryRanges$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryRanges-new"></a>}}
\if{latex}{\out{\hypertarget{method-QueryRanges-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryRanges$new(layout = NULL, ranges = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryRanges-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QueryRanges-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryRanges$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryRanges-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QueryRanges-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryRanges$fromJSON(QueryRangesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryRanges-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QueryRanges-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryRanges$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryRanges-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QueryRanges-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryRanges$fromJSONString(QueryRangesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryRanges-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QueryRanges-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryRanges$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
