% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_reader.R
\docType{class}
\name{QueryReader}
\alias{QueryReader}
\title{QueryReader}
\format{
An \code{R6Class} generator object
}
\description{
QueryReader Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{layout}}{\link{Layout} [optional]}

\item{\code{subarray}}{\link{Subarray} [optional]}

\item{\code{readState}}{\link{ReadState} [optional]}

\item{\code{varOffsetsMode}}{character [optional]}

\item{\code{varOffsetsAddExtraElement}}{character [optional]}

\item{\code{varOffsetsBitsize}}{integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QueryReader-new}{\code{QueryReader$new()}}
\item \href{#method-QueryReader-toJSON}{\code{QueryReader$toJSON()}}
\item \href{#method-QueryReader-fromJSON}{\code{QueryReader$fromJSON()}}
\item \href{#method-QueryReader-toJSONString}{\code{QueryReader$toJSONString()}}
\item \href{#method-QueryReader-fromJSONString}{\code{QueryReader$fromJSONString()}}
\item \href{#method-QueryReader-clone}{\code{QueryReader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryReader-new"></a>}}
\if{latex}{\out{\hypertarget{method-QueryReader-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryReader$new(
  layout = NULL,
  subarray = NULL,
  readState = NULL,
  varOffsetsMode = NULL,
  varOffsetsAddExtraElement = NULL,
  varOffsetsBitsize = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryReader-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QueryReader-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryReader$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryReader-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-QueryReader-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryReader$fromJSON(QueryReaderJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryReader-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QueryReader-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryReader$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryReader-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-QueryReader-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryReader$fromJSONString(QueryReaderJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QueryReader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QueryReader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QueryReader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
