% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querystatus.R
\docType{class}
\name{Querystatus}
\alias{Querystatus}
\title{Querystatus}
\format{
An \code{R6Class} generator object
}
\description{
Querystatus Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Querystatus-new}{\code{Querystatus$new()}}
\item \href{#method-Querystatus-toJSON}{\code{Querystatus$toJSON()}}
\item \href{#method-Querystatus-fromJSON}{\code{Querystatus$fromJSON()}}
\item \href{#method-Querystatus-toJSONString}{\code{Querystatus$toJSONString()}}
\item \href{#method-Querystatus-fromJSONString}{\code{Querystatus$fromJSONString()}}
\item \href{#method-Querystatus-clone}{\code{Querystatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Querystatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-Querystatus-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Querystatus$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Querystatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Querystatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Querystatus$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Querystatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Querystatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Querystatus$fromJSON(QuerystatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Querystatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Querystatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Querystatus$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Querystatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Querystatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Querystatus$fromJSONString(QuerystatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Querystatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Querystatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Querystatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
