% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querytype.R
\docType{class}
\name{Querytype}
\alias{Querytype}
\title{Querytype}
\format{
An \code{R6Class} generator object
}
\description{
Querytype Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Querytype-new}{\code{Querytype$new()}}
\item \href{#method-Querytype-toJSON}{\code{Querytype$toJSON()}}
\item \href{#method-Querytype-fromJSON}{\code{Querytype$fromJSON()}}
\item \href{#method-Querytype-toJSONString}{\code{Querytype$toJSONString()}}
\item \href{#method-Querytype-fromJSONString}{\code{Querytype$fromJSONString()}}
\item \href{#method-Querytype-clone}{\code{Querytype$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Querytype-new"></a>}}
\if{latex}{\out{\hypertarget{method-Querytype-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Querytype$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Querytype-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Querytype-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Querytype$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Querytype-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Querytype-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Querytype$fromJSON(QuerytypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Querytype-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Querytype-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Querytype$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Querytype-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Querytype-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Querytype$fromJSONString(QuerytypeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Querytype-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Querytype-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Querytype$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
