% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_state.R
\docType{class}
\name{ReadState}
\alias{ReadState}
\title{ReadState}
\format{
An \code{R6Class} generator object
}
\description{
ReadState Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{initialized}}{character [optional]}

\item{\code{overflowed}}{character [optional]}

\item{\code{unsplittable}}{character [optional]}

\item{\code{subarrayPartitioner}}{\link{SubarrayPartitioner} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReadState-new}{\code{ReadState$new()}}
\item \href{#method-ReadState-toJSON}{\code{ReadState$toJSON()}}
\item \href{#method-ReadState-fromJSON}{\code{ReadState$fromJSON()}}
\item \href{#method-ReadState-toJSONString}{\code{ReadState$toJSONString()}}
\item \href{#method-ReadState-fromJSONString}{\code{ReadState$fromJSONString()}}
\item \href{#method-ReadState-clone}{\code{ReadState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadState-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReadState-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadState$new(
  initialized = NULL,
  overflowed = NULL,
  unsplittable = NULL,
  subarrayPartitioner = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadState-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ReadState-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadState$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadState-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ReadState-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadState$fromJSON(ReadStateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadState-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ReadState-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadState$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadState-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ReadState-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadState$fromJSONString(ReadStateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReadState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReadState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReadState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
