% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_array.R
\name{register_array}
\alias{register_array}
\title{Register an existing array on TileDB Cloud}
\usage{
register_array(
  namespace = NULL,
  array_name,
  uri,
  description = NULL,
  access_credentials_name = NULL
)
}
\arguments{
\item{namespace}{Namespace within TileDB cloud to charge. If this is null, the
logged-in user's username will be used for the namespace.}

\item{array_name}{The name to call the array in TileDB Cloud.}

\item{uri}{The URI of where the array is stored.}

\item{description}{Optional description field for the array.}

\item{access_credentials_name}{Credentials to access the array storage. If omitted,
the logged-in user's default credentials will be used.}
}
\description{
The underlying storage must already exist.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute_sequentially}()},
\code{\link{compute}()},
\code{\link{delayed_args<-}()},
\code{\link{delayed_args}()},
\code{\link{delayed_array_udf}()},
\code{\link{delayed_generic_udf}()},
\code{\link{delayed_sql}()},
\code{\link{delayed}()},
\code{\link{deregister_array}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
