% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_udf.R
\name{register_udf}
\alias{register_udf}
\title{Register a UDF on TileDB Cloud}
\usage{
register_udf(
  namespace = NULL,
  name,
  type,
  func,
  func_text = NULL,
  version = NULL,
  image_name = NULL,
  readme = NULL,
  license_id = NULL,
  license_text = NULL,
  tags = NULL
)
}
\arguments{
\item{namespace}{Namespace for the UDF to be stored in, e.g. \code{mynamespace}. If omitted, defaults to username.}

\item{name}{character Name for the function to be stored under in TileDB Cloud, e.g. \code{myudfname}.}

\item{type}{character One of \code{generic}, \code{single_array}, or \code{multi_array}.}

\item{func}{An R function which takes a dataframe as first argument.}

\item{version}{character Optional version string.}

\item{image_name}{character}

\item{readme}{README text to be displayed in the TileDB Cloud UI.}

\item{license_id}{character See the TileDB Cloud UI for options.}

\item{license_text}{character See the TileDB Cloud UI for options.}

\item{tags}{list(character) Tags to apply to the UDF.}

\item{exec_raw}{character Text to display in the TileDB Cloud UI's Preview tab. If omitted, a full deparse of \code{func} is used. You can set this to something shorter for brevity if you like.}
}
\value{
No return value.
}
\description{
Registers a user-defined function on TileDB Cloud, so that it may be invoked by name later.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
