% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_format.R
\docType{class}
\name{ResultFormat}
\alias{ResultFormat}
\title{ResultFormat}
\format{
An \code{R6Class} generator object
}
\description{
ResultFormat Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ResultFormat-new}{\code{ResultFormat$new()}}
\item \href{#method-ResultFormat-toJSON}{\code{ResultFormat$toJSON()}}
\item \href{#method-ResultFormat-fromJSON}{\code{ResultFormat$fromJSON()}}
\item \href{#method-ResultFormat-toJSONString}{\code{ResultFormat$toJSONString()}}
\item \href{#method-ResultFormat-fromJSONString}{\code{ResultFormat$fromJSONString()}}
\item \href{#method-ResultFormat-clone}{\code{ResultFormat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultFormat-new"></a>}}
\if{latex}{\out{\hypertarget{method-ResultFormat-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultFormat$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultFormat-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResultFormat-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultFormat$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultFormat-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ResultFormat-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultFormat$fromJSON(ResultFormatJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultFormat-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResultFormat-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultFormat$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultFormat-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ResultFormat-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultFormat$fromJSONString(ResultFormatJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ResultFormat-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ResultFormat-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultFormat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
