% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_array.R
\name{share_array}
\alias{share_array}
\title{Share an existing array on TileDB Cloud}
\usage{
share_array(namespace = NULL, array_name, array_sharing)
}
\arguments{
\item{namespace}{Namespace of the array within TileDB cloud. If this is null, the
logged-in user's username will be used for the namespace.}

\item{array_name}{The name to call the array in TileDB Cloud.}

\item{array_sharing}{\link{ArraySharing} object.}
}
\description{
Specify user/organization, read/write, etc.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute_sequentially}()},
\code{\link{compute}()},
\code{\link{delayed_args<-}()},
\code{\link{delayed_args}()},
\code{\link{delayed_array_udf}()},
\code{\link{delayed_generic_udf}()},
\code{\link{delayed_sql}()},
\code{\link{delayed}()},
\code{\link{deregister_array}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_array}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
