% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_parameters.R
\docType{class}
\name{SQLParameters}
\alias{SQLParameters}
\title{SQLParameters}
\format{
An \code{R6Class} generator object
}
\description{
SQLParameters Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{query}}{character [optional]}

\item{\code{output_uri}}{character [optional]}

\item{\code{store_results}}{character [optional]}

\item{\code{dont_download_results}}{character [optional]}

\item{\code{resource_class}}{character [optional]}

\item{\code{result_format}}{\link{ResultFormat} [optional]}

\item{\code{init_commands}}{list( character ) [optional]}

\item{\code{parameters}}{list( object ) [optional]}

\item{\code{task_graph_uuid}}{character [optional]}

\item{\code{client_node_uuid}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SQLParameters-new}{\code{SQLParameters$new()}}
\item \href{#method-SQLParameters-toJSON}{\code{SQLParameters$toJSON()}}
\item \href{#method-SQLParameters-fromJSON}{\code{SQLParameters$fromJSON()}}
\item \href{#method-SQLParameters-toJSONString}{\code{SQLParameters$toJSONString()}}
\item \href{#method-SQLParameters-fromJSONString}{\code{SQLParameters$fromJSONString()}}
\item \href{#method-SQLParameters-clone}{\code{SQLParameters$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLParameters-new"></a>}}
\if{latex}{\out{\hypertarget{method-SQLParameters-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLParameters$new(
  name = NULL,
  query = NULL,
  output_uri = NULL,
  store_results = NULL,
  dont_download_results = NULL,
  resource_class = NULL,
  result_format = NULL,
  init_commands = NULL,
  parameters = NULL,
  task_graph_uuid = NULL,
  client_node_uuid = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLParameters-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SQLParameters-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLParameters$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLParameters-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SQLParameters-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLParameters$fromJSON(SQLParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLParameters-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SQLParameters-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLParameters$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLParameters-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SQLParameters-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLParameters$fromJSONString(SQLParametersJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLParameters-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SQLParameters-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLParameters$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
