% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso_provider.R
\docType{class}
\name{SSOProvider}
\alias{SSOProvider}
\title{SSOProvider}
\format{
An \code{R6Class} generator object
}
\description{
SSOProvider Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SSOProvider-new}{\code{SSOProvider$new()}}
\item \href{#method-SSOProvider-toJSON}{\code{SSOProvider$toJSON()}}
\item \href{#method-SSOProvider-fromJSON}{\code{SSOProvider$fromJSON()}}
\item \href{#method-SSOProvider-toJSONString}{\code{SSOProvider$toJSONString()}}
\item \href{#method-SSOProvider-fromJSONString}{\code{SSOProvider$fromJSONString()}}
\item \href{#method-SSOProvider-clone}{\code{SSOProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSOProvider-new"></a>}}
\if{latex}{\out{\hypertarget{method-SSOProvider-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSOProvider$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSOProvider-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SSOProvider-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSOProvider$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSOProvider-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SSOProvider-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSOProvider$fromJSON(SSOProviderJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSOProvider-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SSOProvider-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSOProvider$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSOProvider-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SSOProvider-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSOProvider$fromJSONString(SSOProviderJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSOProvider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SSOProvider-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSOProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
