% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subarray.R
\docType{class}
\name{Subarray}
\alias{Subarray}
\title{Subarray}
\format{
An \code{R6Class} generator object
}
\description{
Subarray Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{layout}}{\link{Layout} [optional]}

\item{\code{ranges}}{list( \link{SubarrayRanges} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Subarray-new}{\code{Subarray$new()}}
\item \href{#method-Subarray-toJSON}{\code{Subarray$toJSON()}}
\item \href{#method-Subarray-fromJSON}{\code{Subarray$fromJSON()}}
\item \href{#method-Subarray-toJSONString}{\code{Subarray$toJSONString()}}
\item \href{#method-Subarray-fromJSONString}{\code{Subarray$fromJSONString()}}
\item \href{#method-Subarray-clone}{\code{Subarray$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subarray-new"></a>}}
\if{latex}{\out{\hypertarget{method-Subarray-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subarray$new(layout = NULL, ranges = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subarray-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Subarray-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subarray$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subarray-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Subarray-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subarray$fromJSON(SubarrayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subarray-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Subarray-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subarray$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subarray-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Subarray-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subarray$fromJSONString(SubarrayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subarray-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Subarray-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subarray$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
