% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subarray_partitioner_current.R
\docType{class}
\name{SubarrayPartitionerCurrent}
\alias{SubarrayPartitionerCurrent}
\title{SubarrayPartitionerCurrent}
\format{
An \code{R6Class} generator object
}
\description{
SubarrayPartitionerCurrent Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{subarray}}{\link{Subarray} [optional]}

\item{\code{start}}{integer [optional]}

\item{\code{end}}{integer [optional]}

\item{\code{splitMultiRange}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SubarrayPartitionerCurrent-new}{\code{SubarrayPartitionerCurrent$new()}}
\item \href{#method-SubarrayPartitionerCurrent-toJSON}{\code{SubarrayPartitionerCurrent$toJSON()}}
\item \href{#method-SubarrayPartitionerCurrent-fromJSON}{\code{SubarrayPartitionerCurrent$fromJSON()}}
\item \href{#method-SubarrayPartitionerCurrent-toJSONString}{\code{SubarrayPartitionerCurrent$toJSONString()}}
\item \href{#method-SubarrayPartitionerCurrent-fromJSONString}{\code{SubarrayPartitionerCurrent$fromJSONString()}}
\item \href{#method-SubarrayPartitionerCurrent-clone}{\code{SubarrayPartitionerCurrent$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubarrayPartitionerCurrent-new"></a>}}
\if{latex}{\out{\hypertarget{method-SubarrayPartitionerCurrent-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitionerCurrent$new(
  subarray = NULL,
  start = NULL,
  end = NULL,
  splitMultiRange = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubarrayPartitionerCurrent-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SubarrayPartitionerCurrent-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitionerCurrent$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubarrayPartitionerCurrent-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SubarrayPartitionerCurrent-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitionerCurrent$fromJSON(SubarrayPartitionerCurrentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubarrayPartitionerCurrent-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SubarrayPartitionerCurrent-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitionerCurrent$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubarrayPartitionerCurrent-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SubarrayPartitionerCurrent-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitionerCurrent$fromJSONString(SubarrayPartitionerCurrentJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubarrayPartitionerCurrent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SubarrayPartitionerCurrent-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitionerCurrent$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
