% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subarray_partitioner_state.R
\docType{class}
\name{SubarrayPartitionerState}
\alias{SubarrayPartitionerState}
\title{SubarrayPartitionerState}
\format{
An \code{R6Class} generator object
}
\description{
SubarrayPartitionerState Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{start}}{integer [optional]}

\item{\code{end}}{integer [optional]}

\item{\code{singleRange}}{list( \link{Subarray} ) [optional]}

\item{\code{multiRange}}{list( \link{Subarray} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SubarrayPartitionerState-new}{\code{SubarrayPartitionerState$new()}}
\item \href{#method-SubarrayPartitionerState-toJSON}{\code{SubarrayPartitionerState$toJSON()}}
\item \href{#method-SubarrayPartitionerState-fromJSON}{\code{SubarrayPartitionerState$fromJSON()}}
\item \href{#method-SubarrayPartitionerState-toJSONString}{\code{SubarrayPartitionerState$toJSONString()}}
\item \href{#method-SubarrayPartitionerState-fromJSONString}{\code{SubarrayPartitionerState$fromJSONString()}}
\item \href{#method-SubarrayPartitionerState-clone}{\code{SubarrayPartitionerState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubarrayPartitionerState-new"></a>}}
\if{latex}{\out{\hypertarget{method-SubarrayPartitionerState-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitionerState$new(
  start = NULL,
  end = NULL,
  singleRange = NULL,
  multiRange = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubarrayPartitionerState-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SubarrayPartitionerState-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitionerState$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubarrayPartitionerState-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-SubarrayPartitionerState-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitionerState$fromJSON(SubarrayPartitionerStateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubarrayPartitionerState-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SubarrayPartitionerState-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitionerState$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubarrayPartitionerState-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-SubarrayPartitionerState-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitionerState$fromJSONString(SubarrayPartitionerStateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SubarrayPartitionerState-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SubarrayPartitionerState-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SubarrayPartitionerState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
