% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscription.R
\docType{class}
\name{Subscription}
\alias{Subscription}
\title{Subscription}
\format{
An \code{R6Class} generator object
}
\description{
Subscription Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{owner_namespace_uuid}}{character [optional]}

\item{\code{customer_namespace_uuid}}{character [optional]}

\item{\code{pricing}}{list( \link{Pricing} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Subscription-new}{\code{Subscription$new()}}
\item \href{#method-Subscription-toJSON}{\code{Subscription$toJSON()}}
\item \href{#method-Subscription-fromJSON}{\code{Subscription$fromJSON()}}
\item \href{#method-Subscription-toJSONString}{\code{Subscription$toJSONString()}}
\item \href{#method-Subscription-fromJSONString}{\code{Subscription$fromJSONString()}}
\item \href{#method-Subscription-clone}{\code{Subscription$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subscription-new"></a>}}
\if{latex}{\out{\hypertarget{method-Subscription-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subscription$new(
  id = NULL,
  owner_namespace_uuid = NULL,
  customer_namespace_uuid = NULL,
  pricing = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subscription-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Subscription-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subscription$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subscription-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Subscription-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subscription$fromJSON(SubscriptionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subscription-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Subscription-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subscription$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subscription-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Subscription-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subscription$fromJSONString(SubscriptionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subscription-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Subscription-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subscription$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
