% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph_log.R
\docType{class}
\name{TaskGraphLog}
\alias{TaskGraphLog}
\title{TaskGraphLog}
\format{
An \code{R6Class} generator object
}
\description{
TaskGraphLog Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{uuid}}{character [optional]}

\item{\code{namespace}}{character [optional]}

\item{\code{created_by}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{created_at}}{character [optional]}

\item{\code{start_time}}{character [optional]}

\item{\code{end_time}}{character [optional]}

\item{\code{status}}{\link{TaskGraphLogStatus} [optional]}

\item{\code{nodes}}{list( \link{TaskGraphNodeMetadata} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TaskGraphLog$new()}}
\item \href{#method-toJSON}{\code{TaskGraphLog$toJSON()}}
\item \href{#method-fromJSON}{\code{TaskGraphLog$fromJSON()}}
\item \href{#method-toJSONString}{\code{TaskGraphLog$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TaskGraphLog$fromJSONString()}}
\item \href{#method-clone}{\code{TaskGraphLog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLog$new(
  uuid = NULL,
  namespace = NULL,
  created_by = NULL,
  name = NULL,
  created_at = NULL,
  start_time = NULL,
  end_time = NULL,
  status = NULL,
  nodes = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLog$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLog$fromJSON(TaskGraphLogJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLog$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLog$fromJSONString(TaskGraphLogJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
