% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph_logs_api.R
\docType{class}
\name{TaskGraphLogsApi}
\alias{TaskGraphLogsApi}
\title{TaskGraphLogs operations}
\format{
An \code{R6Class} generator object
}
\description{
tiledbcloud.TaskGraphLogs
}
\section{Methods}{

\describe{
\strong{ CreateTaskGraphLog } \emph{  }
Create a task graph log.

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } log \link{TaskGraphLog}
\item \emph{ @returnType } \link{TaskGraphLog} \cr


\item status code : 201 | The task graph was created. The returned TaskGraphLog will include the data the client sent, with the server-defined fields added in. 

\item return type : TaskGraphLog 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetTaskGraphLog } \emph{  }
Fetch information about a single task graph execution. 

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } id character
\item \emph{ @returnType } \link{TaskGraphLog} \cr


\item status code : 200 | Information about the execution of a single task graph.

\item return type : TaskGraphLog 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListTaskGraphLogs } \emph{  }
Fetch the task graph logs of a namespace the user has access to. The returned entries will include only summary data, and will not include information about the individual tasks that were executed. (This information is available when requesting an individual task graph log.) Entries in the response are ordered from newest to oldest. Pagination parameters work as in other API methods; see PaginationMetadata. 

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } created.by character
\item \emph{ @param } search character
\item \emph{ @param } start.time character
\item \emph{ @param } end.time character
\item \emph{ @param } page integer
\item \emph{ @param } per.page integer
\item \emph{ @returnType } \link{TaskGraphLogsData} \cr


\item status code : 200 | The task graph logs that matched the user&#39;s query.

\item return type : TaskGraphLogsData 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ UpdateTaskGraphLog } \emph{  }
Update information about a single task graph execution. 

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } id character
\item \emph{ @param } log \link{TaskGraphLog}


\item status code : 204 | Log entry updated successfully.


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  CreateTaskGraphLog  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace that will own this task graph log.
var.log <- TaskGraphLog$new() # TaskGraphLog | 

api.instance <- TaskGraphLogsApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$CreateTaskGraphLog(var.namespace, var.log)


####################  GetTaskGraphLog  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace that owns this task graph log.
var.id <- 'id_example' # character | The UUID of the task graph log entry.

api.instance <- TaskGraphLogsApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetTaskGraphLog(var.namespace, var.id)


####################  ListTaskGraphLogs  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | Include logs for this namespace.
var.created.by <- 'created.by_example' # character | Include logs from only this user.
var.search <- 'search_example' # character | search string that will look at name.
var.start.time <- 'start.time_example' # character | Include logs created after this time.
var.end.time <- 'end.time_example' # character | Include logs created before this time.
var.page <- 56 # integer | pagination offset
var.per.page <- 56 # integer | pagination limit

api.instance <- TaskGraphLogsApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ListTaskGraphLogs(namespace=var.namespace, created.by=var.created.by, search=var.search, start.time=var.start.time, end.time=var.end.time, page=var.page, per.page=var.per.page)


####################  UpdateTaskGraphLog  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace that owns this task graph log.
var.id <- 'id_example' # character | The UUID of the task graph log entry.
var.log <- TaskGraphLog$new() # TaskGraphLog | Updates to make to the task graph log. The only manual update that a client should need to make to a task graph log is to update its completion status to `succeeded`, `failed`, or `cancelled`. 

api.instance <- TaskGraphLogsApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$UpdateTaskGraphLog(var.namespace, var.id, var.log)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaskGraphLogsApi-new}{\code{TaskGraphLogsApi$new()}}
\item \href{#method-TaskGraphLogsApi-CreateTaskGraphLog}{\code{TaskGraphLogsApi$CreateTaskGraphLog()}}
\item \href{#method-TaskGraphLogsApi-CreateTaskGraphLogWithHttpInfo}{\code{TaskGraphLogsApi$CreateTaskGraphLogWithHttpInfo()}}
\item \href{#method-TaskGraphLogsApi-GetTaskGraphLog}{\code{TaskGraphLogsApi$GetTaskGraphLog()}}
\item \href{#method-TaskGraphLogsApi-GetTaskGraphLogWithHttpInfo}{\code{TaskGraphLogsApi$GetTaskGraphLogWithHttpInfo()}}
\item \href{#method-TaskGraphLogsApi-ListTaskGraphLogs}{\code{TaskGraphLogsApi$ListTaskGraphLogs()}}
\item \href{#method-TaskGraphLogsApi-ListTaskGraphLogsWithHttpInfo}{\code{TaskGraphLogsApi$ListTaskGraphLogsWithHttpInfo()}}
\item \href{#method-TaskGraphLogsApi-UpdateTaskGraphLog}{\code{TaskGraphLogsApi$UpdateTaskGraphLog()}}
\item \href{#method-TaskGraphLogsApi-UpdateTaskGraphLogWithHttpInfo}{\code{TaskGraphLogsApi$UpdateTaskGraphLogWithHttpInfo()}}
\item \href{#method-TaskGraphLogsApi-clone}{\code{TaskGraphLogsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsApi-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsApi-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsApi-CreateTaskGraphLog"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsApi-CreateTaskGraphLog}{}}}
\subsection{Method \code{CreateTaskGraphLog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsApi$CreateTaskGraphLog(namespace, log, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsApi-CreateTaskGraphLogWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsApi-CreateTaskGraphLogWithHttpInfo}{}}}
\subsection{Method \code{CreateTaskGraphLogWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsApi$CreateTaskGraphLogWithHttpInfo(namespace, log, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsApi-GetTaskGraphLog"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsApi-GetTaskGraphLog}{}}}
\subsection{Method \code{GetTaskGraphLog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsApi$GetTaskGraphLog(namespace, id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsApi-GetTaskGraphLogWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsApi-GetTaskGraphLogWithHttpInfo}{}}}
\subsection{Method \code{GetTaskGraphLogWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsApi$GetTaskGraphLogWithHttpInfo(namespace, id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsApi-ListTaskGraphLogs"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsApi-ListTaskGraphLogs}{}}}
\subsection{Method \code{ListTaskGraphLogs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsApi$ListTaskGraphLogs(
  namespace = NULL,
  created.by = NULL,
  search = NULL,
  start.time = NULL,
  end.time = NULL,
  page = NULL,
  per.page = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsApi-ListTaskGraphLogsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsApi-ListTaskGraphLogsWithHttpInfo}{}}}
\subsection{Method \code{ListTaskGraphLogsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsApi$ListTaskGraphLogsWithHttpInfo(
  namespace = NULL,
  created.by = NULL,
  search = NULL,
  start.time = NULL,
  end.time = NULL,
  page = NULL,
  per.page = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsApi-UpdateTaskGraphLog"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsApi-UpdateTaskGraphLog}{}}}
\subsection{Method \code{UpdateTaskGraphLog()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsApi$UpdateTaskGraphLog(namespace, id, log, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsApi-UpdateTaskGraphLogWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsApi-UpdateTaskGraphLogWithHttpInfo}{}}}
\subsection{Method \code{UpdateTaskGraphLogWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsApi$UpdateTaskGraphLogWithHttpInfo(namespace, id, log, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsApi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsApi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
