% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph_logs_data.R
\docType{class}
\name{TaskGraphLogsData}
\alias{TaskGraphLogsData}
\title{TaskGraphLogsData}
\format{
An \code{R6Class} generator object
}
\description{
TaskGraphLogsData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task_graph_logs}}{list( \link{TaskGraphLog} ) [optional]}

\item{\code{pagination_metadata}}{\link{PaginationMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaskGraphLogsData-new}{\code{TaskGraphLogsData$new()}}
\item \href{#method-TaskGraphLogsData-toJSON}{\code{TaskGraphLogsData$toJSON()}}
\item \href{#method-TaskGraphLogsData-fromJSON}{\code{TaskGraphLogsData$fromJSON()}}
\item \href{#method-TaskGraphLogsData-toJSONString}{\code{TaskGraphLogsData$toJSONString()}}
\item \href{#method-TaskGraphLogsData-fromJSONString}{\code{TaskGraphLogsData$fromJSONString()}}
\item \href{#method-TaskGraphLogsData-clone}{\code{TaskGraphLogsData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsData-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsData$new(task_graph_logs = NULL, pagination_metadata = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsData$fromJSON(TaskGraphLogsDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsData$fromJSONString(TaskGraphLogsDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogsData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogsData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogsData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
