% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph_log_status.R
\docType{class}
\name{TaskGraphLogStatus}
\alias{TaskGraphLogStatus}
\title{TaskGraphLogStatus}
\format{
An \code{R6Class} generator object
}
\description{
TaskGraphLogStatus Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaskGraphLogStatus-new}{\code{TaskGraphLogStatus$new()}}
\item \href{#method-TaskGraphLogStatus-toJSON}{\code{TaskGraphLogStatus$toJSON()}}
\item \href{#method-TaskGraphLogStatus-fromJSON}{\code{TaskGraphLogStatus$fromJSON()}}
\item \href{#method-TaskGraphLogStatus-toJSONString}{\code{TaskGraphLogStatus$toJSONString()}}
\item \href{#method-TaskGraphLogStatus-fromJSONString}{\code{TaskGraphLogStatus$fromJSONString()}}
\item \href{#method-TaskGraphLogStatus-clone}{\code{TaskGraphLogStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogStatus-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogStatus-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogStatus$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogStatus-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogStatus-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogStatus$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogStatus-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogStatus-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogStatus$fromJSON(TaskGraphLogStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogStatus-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogStatus-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogStatus$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogStatus-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogStatus-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogStatus$fromJSONString(TaskGraphLogStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGraphLogStatus-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGraphLogStatus-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphLogStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
