% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_db_config.R
\docType{class}
\name{TileDBConfig}
\alias{TileDBConfig}
\title{TileDBConfig}
\format{
An \code{R6Class} generator object
}
\description{
TileDBConfig Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{configs}}{named list( character ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TileDBConfig-new}{\code{TileDBConfig$new()}}
\item \href{#method-TileDBConfig-toJSON}{\code{TileDBConfig$toJSON()}}
\item \href{#method-TileDBConfig-fromJSON}{\code{TileDBConfig$fromJSON()}}
\item \href{#method-TileDBConfig-toJSONString}{\code{TileDBConfig$toJSONString()}}
\item \href{#method-TileDBConfig-fromJSONString}{\code{TileDBConfig$fromJSONString()}}
\item \href{#method-TileDBConfig-clone}{\code{TileDBConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBConfig-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBConfig$new(configs = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBConfig-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBConfig-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBConfig$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBConfig-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBConfig-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBConfig$fromJSON(TileDBConfigJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBConfig-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBConfig-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBConfig$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBConfig-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBConfig-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBConfig$fromJSONString(TileDBConfigJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
