% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\docType{class}
\name{Token}
\alias{Token}
\title{Token}
\format{
An \code{R6Class} generator object
}
\description{
Token Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{token}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{issued_at}}{character [optional]}

\item{\code{expires_at}}{character [optional]}

\item{\code{scope}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Token-new}{\code{Token$new()}}
\item \href{#method-Token-toJSON}{\code{Token$toJSON()}}
\item \href{#method-Token-fromJSON}{\code{Token$fromJSON()}}
\item \href{#method-Token-toJSONString}{\code{Token$toJSONString()}}
\item \href{#method-Token-fromJSONString}{\code{Token$fromJSONString()}}
\item \href{#method-Token-clone}{\code{Token$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Token-new"></a>}}
\if{latex}{\out{\hypertarget{method-Token-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Token$new(
  token = NULL,
  name = NULL,
  issued_at = NULL,
  expires_at = NULL,
  scope = "*",
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Token-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Token-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Token$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Token-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Token-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Token$fromJSON(TokenJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Token-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Token-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Token$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Token-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Token-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Token$fromJSONString(TokenJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Token-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Token-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Token$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
