% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf_actions.R
\docType{class}
\name{UDFActions}
\alias{UDFActions}
\title{UDFActions}
\format{
An \code{R6Class} generator object
}
\description{
UDFActions Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UDFActions-new}{\code{UDFActions$new()}}
\item \href{#method-UDFActions-toJSON}{\code{UDFActions$toJSON()}}
\item \href{#method-UDFActions-fromJSON}{\code{UDFActions$fromJSON()}}
\item \href{#method-UDFActions-toJSONString}{\code{UDFActions$toJSONString()}}
\item \href{#method-UDFActions-fromJSONString}{\code{UDFActions$fromJSONString()}}
\item \href{#method-UDFActions-clone}{\code{UDFActions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFActions-new"></a>}}
\if{latex}{\out{\hypertarget{method-UDFActions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFActions$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFActions-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFActions-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFActions$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFActions-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFActions-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFActions$fromJSON(UDFActionsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFActions-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFActions-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFActions$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFActions-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFActions-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFActions$fromJSONString(UDFActionsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFActions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UDFActions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFActions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
