% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf_favorites_data.R
\docType{class}
\name{UDFFavoritesData}
\alias{UDFFavoritesData}
\title{UDFFavoritesData}
\format{
An \code{R6Class} generator object
}
\description{
UDFFavoritesData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{udfs}}{list( \link{ArrayInfo} ) [optional]}

\item{\code{pagination_metadata}}{\link{PaginationMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UDFFavoritesData-new}{\code{UDFFavoritesData$new()}}
\item \href{#method-UDFFavoritesData-toJSON}{\code{UDFFavoritesData$toJSON()}}
\item \href{#method-UDFFavoritesData-fromJSON}{\code{UDFFavoritesData$fromJSON()}}
\item \href{#method-UDFFavoritesData-toJSONString}{\code{UDFFavoritesData$toJSONString()}}
\item \href{#method-UDFFavoritesData-fromJSONString}{\code{UDFFavoritesData$fromJSONString()}}
\item \href{#method-UDFFavoritesData-clone}{\code{UDFFavoritesData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFFavoritesData-new"></a>}}
\if{latex}{\out{\hypertarget{method-UDFFavoritesData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFFavoritesData$new(udfs = NULL, pagination_metadata = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFFavoritesData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFFavoritesData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFFavoritesData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFFavoritesData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFFavoritesData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFFavoritesData$fromJSON(UDFFavoritesDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFFavoritesData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFFavoritesData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFFavoritesData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFFavoritesData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFFavoritesData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFFavoritesData$fromJSONString(UDFFavoritesDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFFavoritesData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UDFFavoritesData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFFavoritesData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
