% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf_image.R
\docType{class}
\name{UDFImage}
\alias{UDFImage}
\title{UDFImage}
\format{
An \code{R6Class} generator object
}
\description{
UDFImage Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{language}}{\link{UDFLanguage} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UDFImage-new}{\code{UDFImage$new()}}
\item \href{#method-UDFImage-toJSON}{\code{UDFImage$toJSON()}}
\item \href{#method-UDFImage-fromJSON}{\code{UDFImage$fromJSON()}}
\item \href{#method-UDFImage-toJSONString}{\code{UDFImage$toJSONString()}}
\item \href{#method-UDFImage-fromJSONString}{\code{UDFImage$fromJSONString()}}
\item \href{#method-UDFImage-clone}{\code{UDFImage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFImage-new"></a>}}
\if{latex}{\out{\hypertarget{method-UDFImage-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFImage$new(id = NULL, name = NULL, language = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFImage-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFImage-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFImage$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFImage-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFImage-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFImage$fromJSON(UDFImageJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFImage-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFImage-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFImage$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFImage-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFImage-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFImage$fromJSONString(UDFImageJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFImage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UDFImage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFImage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
