% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf_info.R
\docType{class}
\name{UDFInfo}
\alias{UDFInfo}
\title{UDFInfo}
\format{
An \code{R6Class} generator object
}
\description{
UDFInfo Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{language}}{\link{UDFLanguage} [optional]}

\item{\code{type}}{\link{UDFType} [optional]}

\item{\code{readme}}{character [optional]}

\item{\code{license_id}}{character [optional]}

\item{\code{license_text}}{character [optional]}

\item{\code{tags}}{list( character ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UDFInfo-new}{\code{UDFInfo$new()}}
\item \href{#method-UDFInfo-toJSON}{\code{UDFInfo$toJSON()}}
\item \href{#method-UDFInfo-fromJSON}{\code{UDFInfo$fromJSON()}}
\item \href{#method-UDFInfo-toJSONString}{\code{UDFInfo$toJSONString()}}
\item \href{#method-UDFInfo-fromJSONString}{\code{UDFInfo$fromJSONString()}}
\item \href{#method-UDFInfo-clone}{\code{UDFInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-UDFInfo-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfo$new(
  id = NULL,
  name = NULL,
  language = NULL,
  type = NULL,
  readme = NULL,
  license_id = NULL,
  license_text = NULL,
  tags = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFInfo-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFInfo-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfo$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFInfo-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFInfo-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfo$fromJSON(UDFInfoJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFInfo-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFInfo-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfo$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFInfo-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFInfo-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfo$fromJSONString(UDFInfoJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UDFInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
