% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf_info_update.R
\docType{class}
\name{UDFInfoUpdate}
\alias{UDFInfoUpdate}
\title{UDFInfoUpdate}
\format{
An \code{R6Class} generator object
}
\description{
UDFInfoUpdate Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{character [optional]}

\item{\code{language}}{\link{UDFLanguage} [optional]}

\item{\code{version}}{character [optional]}

\item{\code{image_name}}{character [optional]}

\item{\code{type}}{\link{UDFType} [optional]}

\item{\code{exec}}{character [optional]}

\item{\code{exec_raw}}{character [optional]}

\item{\code{readme}}{character [optional]}

\item{\code{license_id}}{character [optional]}

\item{\code{license_text}}{character [optional]}

\item{\code{tags}}{list( character ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UDFInfoUpdate-new}{\code{UDFInfoUpdate$new()}}
\item \href{#method-UDFInfoUpdate-toJSON}{\code{UDFInfoUpdate$toJSON()}}
\item \href{#method-UDFInfoUpdate-fromJSON}{\code{UDFInfoUpdate$fromJSON()}}
\item \href{#method-UDFInfoUpdate-toJSONString}{\code{UDFInfoUpdate$toJSONString()}}
\item \href{#method-UDFInfoUpdate-fromJSONString}{\code{UDFInfoUpdate$fromJSONString()}}
\item \href{#method-UDFInfoUpdate-clone}{\code{UDFInfoUpdate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFInfoUpdate-new"></a>}}
\if{latex}{\out{\hypertarget{method-UDFInfoUpdate-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoUpdate$new(
  name = NULL,
  language = NULL,
  version = NULL,
  image_name = NULL,
  type = NULL,
  exec = NULL,
  exec_raw = NULL,
  readme = NULL,
  license_id = NULL,
  license_text = NULL,
  tags = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFInfoUpdate-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFInfoUpdate-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoUpdate$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFInfoUpdate-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFInfoUpdate-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoUpdate$fromJSON(UDFInfoUpdateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFInfoUpdate-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFInfoUpdate-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoUpdate$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFInfoUpdate-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFInfoUpdate-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoUpdate$fromJSONString(UDFInfoUpdateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFInfoUpdate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UDFInfoUpdate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFInfoUpdate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
