% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf_language.R
\docType{class}
\name{UDFLanguage}
\alias{UDFLanguage}
\title{UDFLanguage}
\format{
An \code{R6Class} generator object
}
\description{
UDFLanguage Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UDFLanguage-new}{\code{UDFLanguage$new()}}
\item \href{#method-UDFLanguage-toJSON}{\code{UDFLanguage$toJSON()}}
\item \href{#method-UDFLanguage-fromJSON}{\code{UDFLanguage$fromJSON()}}
\item \href{#method-UDFLanguage-toJSONString}{\code{UDFLanguage$toJSONString()}}
\item \href{#method-UDFLanguage-fromJSONString}{\code{UDFLanguage$fromJSONString()}}
\item \href{#method-UDFLanguage-clone}{\code{UDFLanguage$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFLanguage-new"></a>}}
\if{latex}{\out{\hypertarget{method-UDFLanguage-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFLanguage$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFLanguage-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFLanguage-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFLanguage$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFLanguage-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFLanguage-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFLanguage$fromJSON(UDFLanguageJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFLanguage-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFLanguage-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFLanguage$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFLanguage-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFLanguage-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFLanguage$fromJSONString(UDFLanguageJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFLanguage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UDFLanguage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFLanguage$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
