% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udf_subarray.R
\docType{class}
\name{UDFSubarray}
\alias{UDFSubarray}
\title{UDFSubarray}
\format{
An \code{R6Class} generator object
}
\description{
UDFSubarray Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{layout}}{\link{Layout} [optional]}

\item{\code{ranges}}{list( \link{UDFSubarrayRange} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UDFSubarray-new}{\code{UDFSubarray$new()}}
\item \href{#method-UDFSubarray-toJSON}{\code{UDFSubarray$toJSON()}}
\item \href{#method-UDFSubarray-fromJSON}{\code{UDFSubarray$fromJSON()}}
\item \href{#method-UDFSubarray-toJSONString}{\code{UDFSubarray$toJSONString()}}
\item \href{#method-UDFSubarray-fromJSONString}{\code{UDFSubarray$fromJSONString()}}
\item \href{#method-UDFSubarray-clone}{\code{UDFSubarray$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFSubarray-new"></a>}}
\if{latex}{\out{\hypertarget{method-UDFSubarray-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFSubarray$new(layout = NULL, ranges = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFSubarray-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFSubarray-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFSubarray$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFSubarray-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-UDFSubarray-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFSubarray$fromJSON(UDFSubarrayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFSubarray-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFSubarray-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFSubarray$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFSubarray-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-UDFSubarray-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFSubarray$fromJSONString(UDFSubarrayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UDFSubarray-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UDFSubarray-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UDFSubarray$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
