% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_udf.R
\name{unregister_udf}
\alias{unregister_udf}
\title{Unregister a UDF from TileDB Cloud}
\usage{
unregister_udf(name, namespace)
}
\arguments{
\item{name}{Name of the UDF in TileDB Cloud.}

\item{namespace}{Namespace for the UDF in TileDB Cloud.}
}
\value{
No return value.
}
\description{
Deletes a registered UDF. This removes all sharing and cannot be undone.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
