% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_user.R
\name{user_profile}
\alias{user_profile}
\title{TileDB Cloud Manual Layer for User Info}
\usage{
user_profile(include_logo = FALSE)
}
\arguments{
\item{include_logo}{If set to True, include the \code{logo} field in the return value.
By default this is omitted since it's a long base64-encoded string which takes up
a lot of screen space and is likely uninteresting.}
}
\value{
A list of user properties from the currently logged-in TileDB cloud account.
}
\description{
This function shows user information for the  currently logged-in account on
TileDB Cloud.  Nominally you will first call login(); if not, the results of
the last login at ~/.tiledb/cloud.json will be used.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{.wrap_as_api_response}()},
\code{\link{array_info}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{list_arrays}()},
\code{\link{login}()}
}
\concept{{manual-layer functions}}
