% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{wrap_as_api_response}
\alias{wrap_as_api_response}
\title{TileDB Cloud HTTP-response helper}
\usage{
wrap_as_api_response(resp)
}
\arguments{
\item{resp}{A \code{response} S3 object e.g. from \code{httr:GET}.}
}
\value{
An object of type \link{ApiResponse}. In the success case (HTTP 2xx)
its \code{content} slot is the raw HTTP body. In the failure case (otherwise)
its \code{content} slot is error-text from the server.
}
\description{
Used for overriding the openapi autogenerated HTTP-response handling:
* Makes the handling compatible with TileDB REST-server response format
* Surfaces error messages back to the user
Not intended to be exported from this package; for package-internal use only.
}
