% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\docType{class}
\name{Writer}
\alias{Writer}
\title{Writer}
\format{
An \code{R6Class} generator object
}
\description{
Writer Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{checkCoordDups}}{character [optional]}

\item{\code{checkCoordOOB}}{character [optional]}

\item{\code{dedupCoords}}{character [optional]}

\item{\code{subarray}}{\link{DomainArray} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Writer-new}{\code{Writer$new()}}
\item \href{#method-Writer-toJSON}{\code{Writer$toJSON()}}
\item \href{#method-Writer-fromJSON}{\code{Writer$fromJSON()}}
\item \href{#method-Writer-toJSONString}{\code{Writer$toJSONString()}}
\item \href{#method-Writer-fromJSONString}{\code{Writer$fromJSONString()}}
\item \href{#method-Writer-clone}{\code{Writer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Writer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Writer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Writer$new(
  checkCoordDups = NULL,
  checkCoordOOB = NULL,
  dedupCoords = NULL,
  subarray = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Writer-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Writer-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Writer$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Writer-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-Writer-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Writer$fromJSON(WriterJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Writer-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Writer-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Writer$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Writer-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-Writer-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Writer$fromJSONString(WriterJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Writer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Writer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Writer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
