% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_array.R
\name{array_info}
\alias{array_info}
\title{TileDB Cloud Manual Layer for Array Info}
\usage{
array_info(namespace, arrayname)
}
\arguments{
\item{namespace}{Like "TileDB-Inc"}

\item{arrayname}{Like "quickstart_dense"}
}
\value{
A list of array properties
}
\description{
This function shows array information on TileDB Cloud.  Nominally you will
first call login(); if not, the results of the last login at
~/.tiledb/cloud.json will be used.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{.wrap_as_api_response}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
