% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_end_timestamp_data.R
\docType{class}
\name{ArrayEndTimestampData}
\alias{ArrayEndTimestampData}
\title{ArrayEndTimestampData}
\format{
An \code{R6Class} generator object
}
\description{
ArrayEndTimestampData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{end_timestamps}}{list( integer ) [optional]}

\item{\code{pagination_metadata}}{\link{PaginationMetadata} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ArrayEndTimestampData-new}{\code{ArrayEndTimestampData$new()}}
\item \href{#method-ArrayEndTimestampData-toJSON}{\code{ArrayEndTimestampData$toJSON()}}
\item \href{#method-ArrayEndTimestampData-fromJSON}{\code{ArrayEndTimestampData$fromJSON()}}
\item \href{#method-ArrayEndTimestampData-toJSONString}{\code{ArrayEndTimestampData$toJSONString()}}
\item \href{#method-ArrayEndTimestampData-fromJSONString}{\code{ArrayEndTimestampData$fromJSONString()}}
\item \href{#method-ArrayEndTimestampData-clone}{\code{ArrayEndTimestampData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayEndTimestampData-new"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayEndTimestampData-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayEndTimestampData$new(
  end_timestamps = NULL,
  pagination_metadata = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayEndTimestampData-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayEndTimestampData-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayEndTimestampData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayEndTimestampData-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayEndTimestampData-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayEndTimestampData$fromJSON(ArrayEndTimestampDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayEndTimestampData-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayEndTimestampData-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayEndTimestampData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayEndTimestampData-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayEndTimestampData-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayEndTimestampData$fromJSONString(ArrayEndTimestampDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ArrayEndTimestampData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ArrayEndTimestampData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ArrayEndTimestampData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
