% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_node.R
\name{compute}
\alias{compute}
\title{Launch a task graph from a given terminal node in the task graph}
\usage{
compute(
  node,
  timeout_seconds = NULL,
  verbose = FALSE,
  namespace = NULL,
  force_all_local = FALSE
)
}
\arguments{
\item{node}{The object whose args are being set -- nominally, produced by
\code{delayed}, \code{delayed_generic_udf}, etc.}

\item{timeout_seconds}{Number of seconds after which to stop waiting for results.
Note that in-flight computationsa are not cancelled; this is not supported by the
underlying R package we use for concurrency.}

\item{verbose}{If supplied, show the DAG state at the start and end, along with all node start/end.
Also shown are any stdout prints from the individual nodes, but these are only visible once the
compute node has completed.}

\item{namespace}{The namespace to charge for any cloud costs during the execution of the
task graph. This can be null only when all nodes have \code{local}, or when \code{compute}
is called with \code{force_all_local}.}

\item{force_all_local}{While individual nodes can be marked with \code{local=TRUE}
to not be executed on TileDB cloud, this flag overrides the default \code{local=FALSE}
for *all* nodes in the task graph.}
}
\value{
The value of the computation.
}
\description{
The task graph is implicitly defined by various \code{delayed} objects having others
in their argument lists.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute_sequentially}()},
\code{\link{delayed_args<-}()},
\code{\link{delayed_args}()},
\code{\link{delayed_array_udf}()},
\code{\link{delayed_generic_udf}()},
\code{\link{delayed_sql}()},
\code{\link{delayed}()},
\code{\link{deregister_array}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_array}()},
\code{\link{register_udf}()},
\code{\link{share_array}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
