% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_delayed.R
\name{delayed}
\alias{delayed}
\title{delayed is an exported factory function for the "Node" implementation class.}
\usage{
delayed(func, args = NULL, display_name = NULL, do_local = FALSE)
}
\arguments{
\item{display_name}{Optional -- e.g. \code{a} or \code{b}. If omitted, it defaults to a UUID.}

\item{do_local}{If true, execute the functions on the local host; if else, execute them as
UDFs in TileDB Cloud.}
}
\description{
Intentionally, Node has non-public consructor, but a few public accessors.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
