% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_delayed.R
\name{delayed}
\alias{delayed}
\title{Define a function to be executed within a task graph}
\usage{
delayed(func, args = NULL, name = NULL, namespace = NULL, local = FALSE)
}
\arguments{
\item{name}{Optional -- e.g. \code{a} or \code{b}. If omitted, it defaults to a UUID.}

\item{namespace}{If supplied, a namespace to use for executing this particular node.
If omitted, a namespace can be applied at your top-level call to \code{compute}.
If omitted there as well, your logged-in account's default namespace will be used.}

\item{local}{If true, execute the functions on the local host; if else, execute them as
UDFs in TileDB Cloud.}
}
\value{
A task-graph node object on which you can later call \code{compute}.
}
\description{
Define a function to be executed within a task graph
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute_sequentially}()},
\code{\link{compute}()},
\code{\link{delayed_args<-}()},
\code{\link{delayed_args}()},
\code{\link{delayed_array_udf}()},
\code{\link{delayed_generic_udf}()},
\code{\link{delayed_sql}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
