% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_udf.R
\name{execute_generic_udf}
\alias{execute_generic_udf}
\title{Execute a generic UDF on TileDB Cloud}
\usage{
execute_generic_udf(
  udf = NULL,
  registered_udf_name = NULL,
  args = NULL,
  result_format = "native",
  args_format = "native",
  namespace,
  language = "r"
)
}
\arguments{
\item{udf}{An R function. Arguments are specified separately via \code{args}.
One of \code{udf} and \code{registered_udf_name} must be non-null.}

\item{registered_udf_name}{Name of a registered UDF, of the form \code{namespace/udfname}.
Arguments are specified separately via \code{args}.  One of \code{udf} and
\code{registered_udf_name} must be non-null.}

\item{args}{Arguments to the function. If the function takes
no arguments, this can be omitted. If you want to call by
position, use a list like \code{args=list(123, 456)}. If you want
to call by name, use a named list like \code{args=list(x=123,y=456)}.}

\item{result_format}{One of \code{native}, \code{json}, or \code{arrow}. These are
used as wire format for returning results from the server to this library, primarily
for memory-usage control.  UDF return values handed back to your code from this
library are converted back to natural R objects.}

\item{args_format}{One of \code{native}, \code{json}, or \code{arrow}. These are
used as wire format for sending arguments to the server. Normally you do not need
to specify this. If you're invoking an R UDF, \code{native} is used; if you're
invoking a registered Python UDF, \code{json} is used but you can select \code{arrow}
if you wish.}

\item{namespace}{Namespace within TileDB cloud.}

\item{language}{If omitted, defaults to \code{"r"}. Can be set to \code{"python"}}
}
\value{
The R object which is the return value from the UDF.
}
\description{
This invokes a user-defined function in TileDB Cloud.
}
\details{
Nominally you will first call \code{\link{login}}; if not, the results
of the last login at \code{~/.tiledb/cloud.json} will be used.

The \code{udf} and \code{namespace} arguments are required; the \code{args}
argument is optional.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute_sequentially}()},
\code{\link{compute}()},
\code{\link{delayed_args<-}()},
\code{\link{delayed_args}()},
\code{\link{delayed_array_udf}()},
\code{\link{delayed_generic_udf}()},
\code{\link{delayed_sql}()},
\code{\link{delayed}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
