% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_udf.R
\name{execute_generic_udf}
\alias{execute_generic_udf}
\title{Execute a generic UDF on TileDB Cloud}
\usage{
execute_generic_udf(
  namespace,
  udf = NULL,
  registered_udf_name = NULL,
  args = NULL
)
}
\arguments{
\item{namespace}{Namespace within TileDB cloud.}

\item{udf}{An R function. Arguments are specified separately via \code{args}.
One of \code{udf} and \code{registered_udf_name} must be non-null.}

\item{registered_udf_name}{Name of a registered UDF, of the form \code{namespace/udfname}.
Arguments are specified separately via \code{args}.  One of \code{udf} and
\code{registered_udf_name} must be non-null.}

\item{args}{Arguments to the function. If the function takes
no arguments, this can be omitted. If you want to call by
position, use a list like \code{args=list(123, 456)}. If you want
to call by name, use a named list like \code{args=list(x=123,y=456)}.}
}
\value{
The R object which is the return value from the UDF.
}
\description{
This invokes a user-defined function in TileDB Cloud.
}
\details{
Nominally you will first call \code{\link{login}}; if not, the results
of the last login at \code{~/.tiledb/cloud.json} will be used.

The \code{udf} and \code{namespace} arguments are required; the \code{args}
argument is optional.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute}()},
\code{\link{delayedArrayUDF}()},
\code{\link{delayedSQL}()},
\code{\link{delayed}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{execute_sql_query}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()},
\code{\link{user_profile}()}
}
\concept{{manual-layer functions}}
