% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_execute.R
\name{execute_multi_array_udf}
\alias{execute_multi_array_udf}
\title{TileDB Cloud UDF-Execution Helper for multiple-array UDFs}
\usage{
execute_multi_array_udf(
  namespace,
  array_list,
  udf,
  args = NULL,
  result_format = "native"
)
}
\arguments{
\item{namespace}{Namespace within TileDB cloud.}

\item{array_list}{List of \code{UDFArrayDetails} objects.
Example list element: \code{tiledbcloud::UDFArrayDetails$new(uri="tiledb://demo/quickstart_dense", ranges=QueryRanges$new(layout=Layout$new('row-major'), ranges=list(cbind(1,4),cbind(1,4))), buffers=list("a"))}}

\item{udf}{An R function which takes dataframes as arguments, one dataframe argument for each element in \code{array_list}.}

\item{args}{Arguments to the function. If the function takes
no arguments, this can be omitted. If you want to call by
position, use a list like \code{args=list(123, 456)}. If you want
to call by name, use a named list like \code{args=list(x=123,y=456)}.}

\item{result_format}{One of \code{native}, \code{json}, or \code{arrow}. These are
used as wire format for returning results from the server to this library, primarily
for memory-usage control.  UDF return values handed back to your code from this
library are converted back to natural R objects.}
}
\value{
Return value from the UDF.
}
\description{
This function invokes a user-defined function in TileDB Cloud.
Nominally you will first call login(); if not, the results
of the last login at ~/.tiledb/cloud.json will be used.
}
\details{
All arguments are required.
}
