% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorites_api.R
\docType{class}
\name{FavoritesApi}
\alias{FavoritesApi}
\title{Favorites operations}
\format{
An \code{R6Class} generator object
}
\description{
tiledbcloud.Favorites
}
\section{Methods}{

\describe{
\strong{ DeleteFavoriteId } \emph{  }
Delete specific favorite

\itemize{
\item \emph{ @param } id character


\item status code : 200 | OK


\item response headers :

\tabular{ll}{
}
}

\strong{ GetFavorite } \emph{  }
Fetch specific favorite of a user

\itemize{
\item \emph{ @param } id character
\item \emph{ @returnType } \link{Favorite} \cr


\item status code : 200 | OK

\item return type : Favorite 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetFavorites } \emph{  }
Fetch all favorites of connected user

\itemize{
\item \emph{ @param } type character
\item \emph{ @returnType } list( \link{Favorite} ) \cr


\item status code : 200 | OK

\item return type : array[Favorite] 
\item response headers :

\tabular{ll}{
}
}

\strong{ PostFavorite } \emph{  }
Add a new favorite

\itemize{
\item \emph{ @param } body \link{FavoriteUpdate}


\item status code : 200 | OK


\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  DeleteFavoriteId  ####################

library(tiledbcloud)
var.id <- 'id_example' # character | The uuid of the favorite

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$DeleteFavoriteId(var.id)


####################  GetFavorite  ####################

library(tiledbcloud)
var.id <- 'id_example' # character | The uuid of the favorite

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetFavorite(var.id)


####################  GetFavorites  ####################

library(tiledbcloud)
var.type <- 'type_example' # character | The type of the desired favorites e.g. arrays

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetFavorites(type=var.type)


####################  PostFavorite  ####################

library(tiledbcloud)
var.body <- FavoriteUpdate$new() # FavoriteUpdate | 

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$PostFavorite(body=var.body)


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FavoritesApi$new()}}
\item \href{#method-DeleteFavoriteId}{\code{FavoritesApi$DeleteFavoriteId()}}
\item \href{#method-DeleteFavoriteIdWithHttpInfo}{\code{FavoritesApi$DeleteFavoriteIdWithHttpInfo()}}
\item \href{#method-GetFavorite}{\code{FavoritesApi$GetFavorite()}}
\item \href{#method-GetFavoriteWithHttpInfo}{\code{FavoritesApi$GetFavoriteWithHttpInfo()}}
\item \href{#method-GetFavorites}{\code{FavoritesApi$GetFavorites()}}
\item \href{#method-GetFavoritesWithHttpInfo}{\code{FavoritesApi$GetFavoritesWithHttpInfo()}}
\item \href{#method-PostFavorite}{\code{FavoritesApi$PostFavorite()}}
\item \href{#method-PostFavoriteWithHttpInfo}{\code{FavoritesApi$PostFavoriteWithHttpInfo()}}
\item \href{#method-clone}{\code{FavoritesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteFavoriteId"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteFavoriteId}{}}}
\subsection{Method \code{DeleteFavoriteId()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$DeleteFavoriteId(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteFavoriteIdWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteFavoriteIdWithHttpInfo}{}}}
\subsection{Method \code{DeleteFavoriteIdWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$DeleteFavoriteIdWithHttpInfo(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-GetFavorite}{}}}
\subsection{Method \code{GetFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$GetFavorite(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetFavoriteWithHttpInfo}{}}}
\subsection{Method \code{GetFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$GetFavoriteWithHttpInfo(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFavorites"></a>}}
\if{latex}{\out{\hypertarget{method-GetFavorites}{}}}
\subsection{Method \code{GetFavorites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$GetFavorites(type = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetFavoritesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetFavoritesWithHttpInfo}{}}}
\subsection{Method \code{GetFavoritesWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$GetFavoritesWithHttpInfo(type = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PostFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-PostFavorite}{}}}
\subsection{Method \code{PostFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$PostFavorite(body = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PostFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-PostFavoriteWithHttpInfo}{}}}
\subsection{Method \code{PostFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$PostFavoriteWithHttpInfo(body = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
