% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorites_api.R
\docType{class}
\name{FavoritesApi}
\alias{FavoritesApi}
\title{Favorites operations}
\format{
An \code{R6Class} generator object
}
\description{
tiledbcloud.Favorites
}
\section{Methods}{

\describe{
\strong{ AddArrayFavorite } \emph{  }
Add a new array favorite

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } name character


\item status code : 204 | Item added to favorites successfully


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ AddMLModelFavorite } \emph{  }
Add a new ML model favorite

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } name character


\item status code : 204 | Item added to favorites successfully


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ AddNotebookFavorite } \emph{  }
Add a new notebook favorite

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } name character


\item status code : 204 | Item added to favorites successfully


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ AddUDFFavorite } \emph{  }
Add a new UDF favorite

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } name character


\item status code : 204 | Item added to favorites successfully


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ DeleteArrayFavorite } \emph{  }
Delete specific array favorite

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } name character


\item status code : 204 | array favorite item deleted successfully


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ DeleteMLModelFavorite } \emph{  }
Delete specific ML model favorite

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } name character


\item status code : 204 | ML model favorite item deleted successfully


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ DeleteNotebookFavorite } \emph{  }
Delete specific notebook favorite

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } name character


\item status code : 204 | notebook favorite item deleted successfully


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ DeleteUDFFavorite } \emph{  }
Delete specific UDF favorite

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } name character


\item status code : 204 | UDF favorite item deleted successfully


\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetArrayFavorite } \emph{  }
Fetch array favorite of a specific array

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } name character
\item \emph{ @returnType } \link{ArrayFavorite} \cr


\item status code : 200 | OK

\item return type : ArrayFavorite 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetMLModelFavorite } \emph{  }
Fetch ML model favorite of a specific ML model

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } name character
\item \emph{ @returnType } \link{MLModelFavorite} \cr


\item status code : 200 | OK

\item return type : MLModelFavorite 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetNotebookFavorite } \emph{  }
Fetch notebook favorite of a specific notebook

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } name character
\item \emph{ @returnType } \link{NotebookFavorite} \cr


\item status code : 200 | OK

\item return type : NotebookFavorite 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetUDFFavorite } \emph{  }
Fetch UDF favorite of a specific UDF

\itemize{
\item \emph{ @param } namespace character
\item \emph{ @param } name character
\item \emph{ @returnType } \link{UDFFavorite} \cr


\item status code : 200 | OK

\item return type : UDFFavorite 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListArrayFavorites } \emph{  }
Fetch a page of array favorites of connected user

\itemize{
\item \emph{ @param } page integer
\item \emph{ @param } per.page integer
\item \emph{ @returnType } \link{ArrayFavoritesData} \cr


\item status code : 200 | Available array favorites are returned

\item return type : ArrayFavoritesData 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListArrayFavoritesUUIDs } \emph{  }
Fetch all favorite array uuids of connected user

\itemize{
\item \emph{ @returnType } list( \link{ArrayFavorite} ) \cr


\item status code : 200 | Available favorites array uuids are returned

\item return type : array[ArrayFavorite] 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListMLModelFavorites } \emph{  }
Fetch a page of ML models favorites of connected user

\itemize{
\item \emph{ @param } page integer
\item \emph{ @param } per.page integer
\item \emph{ @returnType } \link{MLModelFavoritesData} \cr


\item status code : 200 | Available ML models favorites are returned

\item return type : MLModelFavoritesData 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListMLModelFavoritesUUIDs } \emph{  }
Fetch all favorite ML models uuids of connected user

\itemize{
\item \emph{ @returnType } list( \link{MLModelFavorite} ) \cr


\item status code : 200 | Available favorites ML model uuids are returned

\item return type : array[MLModelFavorite] 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListNotebookFavorites } \emph{  }
Fetch a page of notebook favorites of connected user

\itemize{
\item \emph{ @param } is.dashboard character
\item \emph{ @param } page integer
\item \emph{ @param } per.page integer
\item \emph{ @returnType } \link{NotebookFavoritesData} \cr


\item status code : 200 | Available notebook favorites are returned

\item return type : NotebookFavoritesData 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListNotebookFavoritesUUIDs } \emph{  }
Fetch all favorite notebook uuids of connected user

\itemize{
\item \emph{ @returnType } list( \link{NotebookFavorite} ) \cr


\item status code : 200 | Available favorites notebook uuids are returned

\item return type : array[NotebookFavorite] 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListUDFFavorites } \emph{  }
Fetch a page of UDF favorites of connected user

\itemize{
\item \emph{ @param } page integer
\item \emph{ @param } per.page integer
\item \emph{ @returnType } \link{UDFFavoritesData} \cr


\item status code : 200 | Available UDF favorites are returned

\item return type : UDFFavoritesData 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListUDFFavoritesUUIDs } \emph{  }
Fetch all favorite UDF uuids of connected user

\itemize{
\item \emph{ @returnType } list( \link{UDFFavorite} ) \cr


\item status code : 200 | Available favorites UDF uuids are returned

\item return type : array[UDFFavorite] 
\item response headers :

\tabular{ll}{
}
\item status code : 0 | error response

\item return type : Error 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  AddArrayFavorite  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace of the array
var.name <- 'name_example' # character | The name of the array

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$AddArrayFavorite(var.namespace, var.name)


####################  AddMLModelFavorite  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace of the ML model
var.name <- 'name_example' # character | The name of the ML model

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$AddMLModelFavorite(var.namespace, var.name)


####################  AddNotebookFavorite  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace of the notebook
var.name <- 'name_example' # character | The name of the notebook

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$AddNotebookFavorite(var.namespace, var.name)


####################  AddUDFFavorite  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace of the UDF
var.name <- 'name_example' # character | The name of the UDF

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$AddUDFFavorite(var.namespace, var.name)


####################  DeleteArrayFavorite  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace of the array
var.name <- 'name_example' # character | The name of the array

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$DeleteArrayFavorite(var.namespace, var.name)


####################  DeleteMLModelFavorite  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace of the ML model
var.name <- 'name_example' # character | The name of the ML model

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$DeleteMLModelFavorite(var.namespace, var.name)


####################  DeleteNotebookFavorite  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace of the notebook
var.name <- 'name_example' # character | The name of the notebook

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$DeleteNotebookFavorite(var.namespace, var.name)


####################  DeleteUDFFavorite  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace of the UDF
var.name <- 'name_example' # character | The name of the UDF

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$DeleteUDFFavorite(var.namespace, var.name)


####################  GetArrayFavorite  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace of the array
var.name <- 'name_example' # character | The name of the array

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetArrayFavorite(var.namespace, var.name)


####################  GetMLModelFavorite  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace of the ML model
var.name <- 'name_example' # character | The name of the ML model

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetMLModelFavorite(var.namespace, var.name)


####################  GetNotebookFavorite  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace of the notebook
var.name <- 'name_example' # character | The name of the notebook

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetNotebookFavorite(var.namespace, var.name)


####################  GetUDFFavorite  ####################

library(tiledbcloud)
var.namespace <- 'namespace_example' # character | The namespace of the UDF
var.name <- 'name_example' # character | The name of the UDF

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$GetUDFFavorite(var.namespace, var.name)


####################  ListArrayFavorites  ####################

library(tiledbcloud)
var.page <- 56 # integer | pagination offset
var.per.page <- 56 # integer | pagination limit

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ListArrayFavorites(page=var.page, per.page=var.per.page)


####################  ListArrayFavoritesUUIDs  ####################

library(tiledbcloud)

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ListArrayFavoritesUUIDs()


####################  ListMLModelFavorites  ####################

library(tiledbcloud)
var.page <- 56 # integer | pagination offset
var.per.page <- 56 # integer | pagination limit

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ListMLModelFavorites(page=var.page, per.page=var.per.page)


####################  ListMLModelFavoritesUUIDs  ####################

library(tiledbcloud)

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ListMLModelFavoritesUUIDs()


####################  ListNotebookFavorites  ####################

library(tiledbcloud)
var.is.dashboard <- 'is.dashboard_example' # character | return only dashboards
var.page <- 56 # integer | pagination offset
var.per.page <- 56 # integer | pagination limit

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ListNotebookFavorites(is.dashboard=var.is.dashboard, page=var.page, per.page=var.per.page)


####################  ListNotebookFavoritesUUIDs  ####################

library(tiledbcloud)

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ListNotebookFavoritesUUIDs()


####################  ListUDFFavorites  ####################

library(tiledbcloud)
var.page <- 56 # integer | pagination offset
var.per.page <- 56 # integer | pagination limit

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ListUDFFavorites(page=var.page, per.page=var.per.page)


####################  ListUDFFavoritesUUIDs  ####################

library(tiledbcloud)

api.instance <- FavoritesApi$new()

#Configure API key authorization: ApiKeyAuth
api.instance$apiClient$apiKeys['X-TILEDB-REST-API-KEY'] <- 'TODO_YOUR_API_KEY';

#Configure HTTP basic authorization: BasicAuth
# provide your username in the user-serial format
api.instance$apiClient$username <- '<user-serial>'; 
# provide your api key generated using the developer portal
api.instance$apiClient$password <- '<api_key>';

result <- api.instance$ListUDFFavoritesUUIDs()


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FavoritesApi$new()}}
\item \href{#method-AddArrayFavorite}{\code{FavoritesApi$AddArrayFavorite()}}
\item \href{#method-AddArrayFavoriteWithHttpInfo}{\code{FavoritesApi$AddArrayFavoriteWithHttpInfo()}}
\item \href{#method-AddMLModelFavorite}{\code{FavoritesApi$AddMLModelFavorite()}}
\item \href{#method-AddMLModelFavoriteWithHttpInfo}{\code{FavoritesApi$AddMLModelFavoriteWithHttpInfo()}}
\item \href{#method-AddNotebookFavorite}{\code{FavoritesApi$AddNotebookFavorite()}}
\item \href{#method-AddNotebookFavoriteWithHttpInfo}{\code{FavoritesApi$AddNotebookFavoriteWithHttpInfo()}}
\item \href{#method-AddUDFFavorite}{\code{FavoritesApi$AddUDFFavorite()}}
\item \href{#method-AddUDFFavoriteWithHttpInfo}{\code{FavoritesApi$AddUDFFavoriteWithHttpInfo()}}
\item \href{#method-DeleteArrayFavorite}{\code{FavoritesApi$DeleteArrayFavorite()}}
\item \href{#method-DeleteArrayFavoriteWithHttpInfo}{\code{FavoritesApi$DeleteArrayFavoriteWithHttpInfo()}}
\item \href{#method-DeleteMLModelFavorite}{\code{FavoritesApi$DeleteMLModelFavorite()}}
\item \href{#method-DeleteMLModelFavoriteWithHttpInfo}{\code{FavoritesApi$DeleteMLModelFavoriteWithHttpInfo()}}
\item \href{#method-DeleteNotebookFavorite}{\code{FavoritesApi$DeleteNotebookFavorite()}}
\item \href{#method-DeleteNotebookFavoriteWithHttpInfo}{\code{FavoritesApi$DeleteNotebookFavoriteWithHttpInfo()}}
\item \href{#method-DeleteUDFFavorite}{\code{FavoritesApi$DeleteUDFFavorite()}}
\item \href{#method-DeleteUDFFavoriteWithHttpInfo}{\code{FavoritesApi$DeleteUDFFavoriteWithHttpInfo()}}
\item \href{#method-GetArrayFavorite}{\code{FavoritesApi$GetArrayFavorite()}}
\item \href{#method-GetArrayFavoriteWithHttpInfo}{\code{FavoritesApi$GetArrayFavoriteWithHttpInfo()}}
\item \href{#method-GetMLModelFavorite}{\code{FavoritesApi$GetMLModelFavorite()}}
\item \href{#method-GetMLModelFavoriteWithHttpInfo}{\code{FavoritesApi$GetMLModelFavoriteWithHttpInfo()}}
\item \href{#method-GetNotebookFavorite}{\code{FavoritesApi$GetNotebookFavorite()}}
\item \href{#method-GetNotebookFavoriteWithHttpInfo}{\code{FavoritesApi$GetNotebookFavoriteWithHttpInfo()}}
\item \href{#method-GetUDFFavorite}{\code{FavoritesApi$GetUDFFavorite()}}
\item \href{#method-GetUDFFavoriteWithHttpInfo}{\code{FavoritesApi$GetUDFFavoriteWithHttpInfo()}}
\item \href{#method-ListArrayFavorites}{\code{FavoritesApi$ListArrayFavorites()}}
\item \href{#method-ListArrayFavoritesWithHttpInfo}{\code{FavoritesApi$ListArrayFavoritesWithHttpInfo()}}
\item \href{#method-ListArrayFavoritesUUIDs}{\code{FavoritesApi$ListArrayFavoritesUUIDs()}}
\item \href{#method-ListArrayFavoritesUUIDsWithHttpInfo}{\code{FavoritesApi$ListArrayFavoritesUUIDsWithHttpInfo()}}
\item \href{#method-ListMLModelFavorites}{\code{FavoritesApi$ListMLModelFavorites()}}
\item \href{#method-ListMLModelFavoritesWithHttpInfo}{\code{FavoritesApi$ListMLModelFavoritesWithHttpInfo()}}
\item \href{#method-ListMLModelFavoritesUUIDs}{\code{FavoritesApi$ListMLModelFavoritesUUIDs()}}
\item \href{#method-ListMLModelFavoritesUUIDsWithHttpInfo}{\code{FavoritesApi$ListMLModelFavoritesUUIDsWithHttpInfo()}}
\item \href{#method-ListNotebookFavorites}{\code{FavoritesApi$ListNotebookFavorites()}}
\item \href{#method-ListNotebookFavoritesWithHttpInfo}{\code{FavoritesApi$ListNotebookFavoritesWithHttpInfo()}}
\item \href{#method-ListNotebookFavoritesUUIDs}{\code{FavoritesApi$ListNotebookFavoritesUUIDs()}}
\item \href{#method-ListNotebookFavoritesUUIDsWithHttpInfo}{\code{FavoritesApi$ListNotebookFavoritesUUIDsWithHttpInfo()}}
\item \href{#method-ListUDFFavorites}{\code{FavoritesApi$ListUDFFavorites()}}
\item \href{#method-ListUDFFavoritesWithHttpInfo}{\code{FavoritesApi$ListUDFFavoritesWithHttpInfo()}}
\item \href{#method-ListUDFFavoritesUUIDs}{\code{FavoritesApi$ListUDFFavoritesUUIDs()}}
\item \href{#method-ListUDFFavoritesUUIDsWithHttpInfo}{\code{FavoritesApi$ListUDFFavoritesUUIDsWithHttpInfo()}}
\item \href{#method-clone}{\code{FavoritesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddArrayFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-AddArrayFavorite}{}}}
\subsection{Method \code{AddArrayFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$AddArrayFavorite(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddArrayFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AddArrayFavoriteWithHttpInfo}{}}}
\subsection{Method \code{AddArrayFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$AddArrayFavoriteWithHttpInfo(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddMLModelFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-AddMLModelFavorite}{}}}
\subsection{Method \code{AddMLModelFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$AddMLModelFavorite(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddMLModelFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AddMLModelFavoriteWithHttpInfo}{}}}
\subsection{Method \code{AddMLModelFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$AddMLModelFavoriteWithHttpInfo(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddNotebookFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-AddNotebookFavorite}{}}}
\subsection{Method \code{AddNotebookFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$AddNotebookFavorite(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddNotebookFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AddNotebookFavoriteWithHttpInfo}{}}}
\subsection{Method \code{AddNotebookFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$AddNotebookFavoriteWithHttpInfo(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddUDFFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-AddUDFFavorite}{}}}
\subsection{Method \code{AddUDFFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$AddUDFFavorite(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddUDFFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AddUDFFavoriteWithHttpInfo}{}}}
\subsection{Method \code{AddUDFFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$AddUDFFavoriteWithHttpInfo(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteArrayFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteArrayFavorite}{}}}
\subsection{Method \code{DeleteArrayFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$DeleteArrayFavorite(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteArrayFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteArrayFavoriteWithHttpInfo}{}}}
\subsection{Method \code{DeleteArrayFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$DeleteArrayFavoriteWithHttpInfo(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteMLModelFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteMLModelFavorite}{}}}
\subsection{Method \code{DeleteMLModelFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$DeleteMLModelFavorite(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteMLModelFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteMLModelFavoriteWithHttpInfo}{}}}
\subsection{Method \code{DeleteMLModelFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$DeleteMLModelFavoriteWithHttpInfo(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteNotebookFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteNotebookFavorite}{}}}
\subsection{Method \code{DeleteNotebookFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$DeleteNotebookFavorite(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteNotebookFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteNotebookFavoriteWithHttpInfo}{}}}
\subsection{Method \code{DeleteNotebookFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$DeleteNotebookFavoriteWithHttpInfo(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteUDFFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteUDFFavorite}{}}}
\subsection{Method \code{DeleteUDFFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$DeleteUDFFavorite(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteUDFFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteUDFFavoriteWithHttpInfo}{}}}
\subsection{Method \code{DeleteUDFFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$DeleteUDFFavoriteWithHttpInfo(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArrayFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-GetArrayFavorite}{}}}
\subsection{Method \code{GetArrayFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$GetArrayFavorite(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetArrayFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetArrayFavoriteWithHttpInfo}{}}}
\subsection{Method \code{GetArrayFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$GetArrayFavoriteWithHttpInfo(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetMLModelFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-GetMLModelFavorite}{}}}
\subsection{Method \code{GetMLModelFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$GetMLModelFavorite(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetMLModelFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetMLModelFavoriteWithHttpInfo}{}}}
\subsection{Method \code{GetMLModelFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$GetMLModelFavoriteWithHttpInfo(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetNotebookFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-GetNotebookFavorite}{}}}
\subsection{Method \code{GetNotebookFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$GetNotebookFavorite(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetNotebookFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetNotebookFavoriteWithHttpInfo}{}}}
\subsection{Method \code{GetNotebookFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$GetNotebookFavoriteWithHttpInfo(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUDFFavorite"></a>}}
\if{latex}{\out{\hypertarget{method-GetUDFFavorite}{}}}
\subsection{Method \code{GetUDFFavorite()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$GetUDFFavorite(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUDFFavoriteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetUDFFavoriteWithHttpInfo}{}}}
\subsection{Method \code{GetUDFFavoriteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$GetUDFFavoriteWithHttpInfo(namespace, name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListArrayFavorites"></a>}}
\if{latex}{\out{\hypertarget{method-ListArrayFavorites}{}}}
\subsection{Method \code{ListArrayFavorites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListArrayFavorites(page = NULL, per.page = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListArrayFavoritesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListArrayFavoritesWithHttpInfo}{}}}
\subsection{Method \code{ListArrayFavoritesWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListArrayFavoritesWithHttpInfo(page = NULL, per.page = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListArrayFavoritesUUIDs"></a>}}
\if{latex}{\out{\hypertarget{method-ListArrayFavoritesUUIDs}{}}}
\subsection{Method \code{ListArrayFavoritesUUIDs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListArrayFavoritesUUIDs(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListArrayFavoritesUUIDsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListArrayFavoritesUUIDsWithHttpInfo}{}}}
\subsection{Method \code{ListArrayFavoritesUUIDsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListArrayFavoritesUUIDsWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMLModelFavorites"></a>}}
\if{latex}{\out{\hypertarget{method-ListMLModelFavorites}{}}}
\subsection{Method \code{ListMLModelFavorites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListMLModelFavorites(page = NULL, per.page = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMLModelFavoritesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListMLModelFavoritesWithHttpInfo}{}}}
\subsection{Method \code{ListMLModelFavoritesWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListMLModelFavoritesWithHttpInfo(
  page = NULL,
  per.page = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMLModelFavoritesUUIDs"></a>}}
\if{latex}{\out{\hypertarget{method-ListMLModelFavoritesUUIDs}{}}}
\subsection{Method \code{ListMLModelFavoritesUUIDs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListMLModelFavoritesUUIDs(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListMLModelFavoritesUUIDsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListMLModelFavoritesUUIDsWithHttpInfo}{}}}
\subsection{Method \code{ListMLModelFavoritesUUIDsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListMLModelFavoritesUUIDsWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListNotebookFavorites"></a>}}
\if{latex}{\out{\hypertarget{method-ListNotebookFavorites}{}}}
\subsection{Method \code{ListNotebookFavorites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListNotebookFavorites(
  is.dashboard = NULL,
  page = NULL,
  per.page = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListNotebookFavoritesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListNotebookFavoritesWithHttpInfo}{}}}
\subsection{Method \code{ListNotebookFavoritesWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListNotebookFavoritesWithHttpInfo(
  is.dashboard = NULL,
  page = NULL,
  per.page = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListNotebookFavoritesUUIDs"></a>}}
\if{latex}{\out{\hypertarget{method-ListNotebookFavoritesUUIDs}{}}}
\subsection{Method \code{ListNotebookFavoritesUUIDs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListNotebookFavoritesUUIDs(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListNotebookFavoritesUUIDsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListNotebookFavoritesUUIDsWithHttpInfo}{}}}
\subsection{Method \code{ListNotebookFavoritesUUIDsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListNotebookFavoritesUUIDsWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUDFFavorites"></a>}}
\if{latex}{\out{\hypertarget{method-ListUDFFavorites}{}}}
\subsection{Method \code{ListUDFFavorites()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListUDFFavorites(page = NULL, per.page = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUDFFavoritesWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListUDFFavoritesWithHttpInfo}{}}}
\subsection{Method \code{ListUDFFavoritesWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListUDFFavoritesWithHttpInfo(page = NULL, per.page = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUDFFavoritesUUIDs"></a>}}
\if{latex}{\out{\hypertarget{method-ListUDFFavoritesUUIDs}{}}}
\subsection{Method \code{ListUDFFavoritesUUIDs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListUDFFavoritesUUIDs(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUDFFavoritesUUIDsWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListUDFFavoritesUUIDsWithHttpInfo}{}}}
\subsection{Method \code{ListUDFFavoritesUUIDsWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$ListUDFFavoritesUUIDsWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FavoritesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
