% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_login.R
\name{get_api_client_instance}
\alias{get_api_client_instance}
\title{TileDB Cloud Login-Reuse Helper}
\usage{
get_api_client_instance()
}
\value{
The cached ApiClient object, or stops.
}
\description{
This is a package-internal function. It returns the cached ApiClient object,
or stops. The ApiClient instance is constructor input to UserApi, ArrayApi, etc.
as generated by OpenAPI.
}
