% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_model_favorite.R
\docType{class}
\name{MLModelFavorite}
\alias{MLModelFavorite}
\title{MLModelFavorite}
\format{
An \code{R6Class} generator object
}
\description{
MLModelFavorite Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mlmodel_uuid}}{character [optional]}

\item{\code{namespace}}{character [optional]}

\item{\code{name}}{character [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLModelFavorite-new}{\code{MLModelFavorite$new()}}
\item \href{#method-MLModelFavorite-toJSON}{\code{MLModelFavorite$toJSON()}}
\item \href{#method-MLModelFavorite-fromJSON}{\code{MLModelFavorite$fromJSON()}}
\item \href{#method-MLModelFavorite-toJSONString}{\code{MLModelFavorite$toJSONString()}}
\item \href{#method-MLModelFavorite-fromJSONString}{\code{MLModelFavorite$fromJSONString()}}
\item \href{#method-MLModelFavorite-clone}{\code{MLModelFavorite$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModelFavorite-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLModelFavorite-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModelFavorite$new(mlmodel_uuid = NULL, namespace = NULL, name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModelFavorite-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MLModelFavorite-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModelFavorite$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModelFavorite-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-MLModelFavorite-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModelFavorite$fromJSON(MLModelFavoriteJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModelFavorite-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MLModelFavorite-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModelFavorite$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModelFavorite-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-MLModelFavorite-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModelFavorite$fromJSONString(MLModelFavoriteJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLModelFavorite-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLModelFavorite-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLModelFavorite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
