% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph_node_metadata.R
\docType{class}
\name{TaskGraphNodeMetadata}
\alias{TaskGraphNodeMetadata}
\title{TaskGraphNodeMetadata}
\format{
An \code{R6Class} generator object
}
\description{
TaskGraphNodeMetadata Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client_node_uuid}}{character [optional]}

\item{\code{name}}{character [optional]}

\item{\code{depends_on}}{list( character ) [optional]}

\item{\code{executions}}{list( \link{ArrayTask} ) [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TaskGraphNodeMetadata$new()}}
\item \href{#method-toJSON}{\code{TaskGraphNodeMetadata$toJSON()}}
\item \href{#method-fromJSON}{\code{TaskGraphNodeMetadata$fromJSON()}}
\item \href{#method-toJSONString}{\code{TaskGraphNodeMetadata$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TaskGraphNodeMetadata$fromJSONString()}}
\item \href{#method-clone}{\code{TaskGraphNodeMetadata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphNodeMetadata$new(
  client_node_uuid = NULL,
  name = NULL,
  depends_on = NULL,
  executions = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphNodeMetadata$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphNodeMetadata$fromJSON(TaskGraphNodeMetadataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphNodeMetadata$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphNodeMetadata$fromJSONString(TaskGraphNodeMetadataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGraphNodeMetadata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
