% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_layer_user.R
\name{user_profile}
\alias{user_profile}
\title{Show information from user TileDB Cloud user profile}
\usage{
user_profile(include_logo = FALSE)
}
\arguments{
\item{include_logo}{If set to True, include the \code{logo} field in the return value.
By default this is omitted since it's a long base64-encoded string which takes up
a lot of screen space and is likely uninteresting.}
}
\value{
A list of user properties from the currently logged-in TileDB cloud account.
}
\description{
This function shows user information for the  currently logged-in account on
TileDB Cloud.
}
\details{
Nominally you will first call \code{\link{login}}; if not, the results of the last
login at \code{~/.tiledb/cloud.json} will be used.
}
\seealso{
Other {manual-layer functions}: 
\code{\link{array_info}()},
\code{\link{compute}()},
\code{\link{delayed}()},
\code{\link{deregister_udf}()},
\code{\link{execute_array_udf}()},
\code{\link{execute_generic_udf}()},
\code{\link{execute_multi_array_udf}()},
\code{\link{get_udf_info}()},
\code{\link{list_arrays}()},
\code{\link{login}()},
\code{\link{register_udf}()},
\code{\link{update_udf_info}()}
}
\concept{{manual-layer functions}}
